### 2022-02-02 #4125 @whedon
[PRE REVIEW]: PyAMG: Algebraic Multigrid Solvers in Python
**Submitting author:** @lukeolson (<a href="http://orcid.org/0000-0002-5283-6104">Luke Olson</a>)
**Repository:** <a href="https://github.com/pyamg/pyamg/" target ="_blank">https://github.com/pyamg/pyamg/</a>
**Version:** v4.2.2
