### 2022-02-02 #4125 @Kevin-Mattheus-Moerman
@openjournals/dev can you see what happened here with the reference check?
