### 2022-02-02 b91711f
docs: fix bib entry
