+++ docs/paper/paper.bib
@@ -96,11 +96,8 @@
 @article{amgintro,
   title   = {An Introduction to Algebraic Multigrid},
   author  = {Falgout, R D},
-  doi     = {},
   url     = {https://www.osti.gov/biblio/897960},
   journal = {Computing in Science and Engineering, vol. 8, no. 6, November 1, 2006, pp. 24-33},
-  number  = ,
-  volume  = ,
   place   = {United States},
   year    = {2006},
   month   = {4}
