### 2022-02-02 bb9cc07
docs: update bib entry
