+++ docs/paper/paper.bib
@@ -95,10 +95,14 @@
 
 @article{amgintro,
   title   = {An Introduction to Algebraic Multigrid},
-  author  = {Falgout, R D},
+  author  = {Falgout, R. D.},
   url     = {https://www.osti.gov/biblio/897960},
-  journal = {Computing in Science and Engineering, vol. 8, no. 6, November 1, 2006, pp. 24-33},
+  journal = {Computing in Science \& Engineering},
+  doi     = {10.1109/MCSE.2006.105},
+  volume  = 8,
+  number  = 6,
+  pages   = {24-33},
   place   = {United States},
-  year    = {2006},
-  month   = {4}
+  year    = 2006,
+  month   = Oct,
 }
