### 2022-02-02 #4125 @lukeolson
> @Kevin-Mattheus-Moerman & @lukeolson - the error says the last bib entry is the problem. I suggest changing it to
>
> ```
> @article{amgintro,
>   title   = {An Introduction to Algebraic Multigrid},
>   author  = {Falgout, R. D.},
>   url     = {https://www.osti.gov/biblio/897960},
>   journal = {Computing in Science \& Engineering},
>   doi = {10.1109/MCSE.2006.105},
>   volume = 8,
>   number = 6,
>   pages = {24-33},
>   place   = {United States},
>   year    = 2006,
>   month   = oct,
> }
> ```

Thanks @danielskatz .  I made this change.
