+++ docs/paper/paper.bib
@@ -57,6 +57,7 @@
   publisher  = {Springer Berlin Heidelberg},
   address    = {Berlin, Heidelberg},
   pages      = {977--981},
+  doi        = {10.1007/978-3-540-70529-1_337},
   url        = {https://lukeo.cs.illinois.edu/files/2015_Ol_encamg.pdf},
 }
 
@@ -70,6 +71,7 @@
   publisher = {Springer Berlin Heidelberg},
   address   = {Berlin, Heidelberg},
   pages     = {981--987},
+  doi       = {10.1007/978-3-540-70529-1_338},
   url       = {https://lukeo.cs.illinois.edu/files/2015_Ol_encmg.pdf},
 }
 
