### 2022-02-02 #4125 @lukeolson
> @lukeolson can you check that missing DOI?

done.  added two additional DOI
