### 2022-02-02 #4125 @danielskatz
@lukeolson - you can use the command `@whedon check references` to check again, and the command `@whedon generate pdf` after making changes to the .md file or when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
