### 2022-02-02 #4125 @danielskatz
> @danielskatz gotcha (I wasn't sure if I was allowed to interact with whedon)

Yes, and encouraged, to make the journal as scalable as possible.  Whedon has some controls, so different roles can do different things, but these commands related to the paper are open to authors.
