### 2022-02-02 2e4f166
feat: add bootstrap and basic test
the test is only for runability at the moment
