+++ pyamg/__init__.py
@@ -13,7 +13,7 @@ from . import (blackbox, graph, graph_ref, multilevel, strength)
 
 from .multilevel import coarse_grid_solver, multilevel_solver, MultilevelSolver
 from .classical import ruge_stuben_solver
-from .aggregation import smoothed_aggregation_solver, rootnode_solver
+from .aggregation import smoothed_aggregation_solver, rootnode_solver, bootstrap_solver
 from .gallery import demo
 from .blackbox import solve, solver, solver_configuration
 
@@ -22,7 +22,8 @@ __all__ = ['__version_tuple__', '__version__',
            'util', 'vis',
            'blackbox', 'graph', 'graph_ref', 'multilevel', 'strength',
            'coarse_grid_solver', 'multilevel_solver', 'MultilevelSolver',
-           'ruge_stuben_solver', 'smoothed_aggregation_solver', 'rootnode_solver',
+           'ruge_stuben_solver', 'smoothed_aggregation_solver',
+           'rootnode_solver', 'bootstrap_solver',
            'demo', 'solve', 'solver', 'solver_configuration']
 
 __all__ += ['test']
