+++ pyamg/aggregation/bootstrap.py
@@ -17,14 +17,14 @@ from .smooth import energy_prolongation_smoother
 from .tentative import fit_candidates
 
 
-def _blocksize(A):
+def blocksize(A):
     """Return the blocksize of a matrix."""
     if isspmatrix_bsr(A):
         return A.blocksize[0]
     return 1
 
 
-def _unpack_arg(v):
+def unpack_arg(v):
     """Unpack arguments for local methods."""
     if isinstance(v, tuple):
         return v[0], v[1]
@@ -55,7 +55,7 @@ def cost(smoother):
     For example, the symmetric option for gauss_seidel results in a multiplier
     that is a factor of 2, and number of iterations is also a multiplier.
     """
-    _, kwargs = _unpack_arg(smoother)
+    _, kwargs = unpack_arg(smoother)
     multiplier = 1.0
     if 'iterations' in kwargs:
         multiplier *= kwargs['iterations']
@@ -95,7 +95,7 @@ def eigen_relaxation_as_linear_operator(A, M, smoother, candidate_iters):
     A - <Ax, x>/<Mx, x> M
     for a homogeneous right-hand-side.
     """
-    fn, kwargs = _unpack_arg(smoother)
+    fn, kwargs = unpack_arg(smoother)
     b = np.zeros((A.shape[0],), dtype=A.dtype)
 
     if fn == 'gauss_seidel':
@@ -163,7 +163,7 @@ def relaxation_as_linear_operator(A, smoother, candidate_iters):
     LinearOperator object that can be used to "multiply" vectors to carry out
     relaxation based on the matrix A for a homogeneous right-hand-side.
     """
-    fn, kwargs = _unpack_arg(smoother)
+    fn, kwargs = unpack_arg(smoother)
     b = np.zeros((A.shape[0],), dtype=A.dtype)
 
     if fn == 'gauss_seidel':
@@ -194,7 +194,7 @@ def relaxation_as_linear_operator(A, smoother, candidate_iters):
     return LinearOperator(A.shape, matvec, dtype=A.dtype)
 
 
-def bootamg_rootnode(A,
+def bootstrap_solver(A,
                      symmetry='hermitian',
                      initial_candidates=None,
                      cands_to_freeze=None,
@@ -211,7 +211,8 @@ def bootamg_rootnode(A,
                      smooth=('energy', {'krylov': 'cg', 'maxiter': 4,
                                         'degree': 2, 'weighting': 'local'}),
                      coarse_solver='pinv2',
-                     prepostsmoother=('gauss_seidel', {'sweep': 'symmetric'})):
+                     prepostsmoother=('gauss_seidel', {'sweep': 'symmetric'}),
+                     verbose=False):
     """Construct a bootstrap AMG solver in a rootnode SA framework.
 
     Parameters
@@ -266,6 +267,8 @@ def bootamg_rootnode(A,
         values like 'gauss_seidel'.  If tuple, use values like ('gauss_seidel',
         {'sweep':'symmetric'}), where the first tuple value is the relaxation
         method, and the second tuple value is a parameter dictionary.
+    verbose : bool
+        Print diagnostic statements in bootstrap_setup
 
     Returns
     -------
@@ -290,10 +293,10 @@ def bootamg_rootnode(A,
     Examples
     --------
     >>> from pyamg.gallery import stencil_grid
-    >>> from bootamg import bootamg_rootnode
+    >>> from pyamg import bootstrap_solver
     >>> import numpy as np
     >>> A=stencil_grid([[-1,-1,-1],[-1,8.0,-1],[-1,-1,-1]], (31,31),format='csr')
-    >>> [asa,work] = bootamg_rootnode(A,num_candidates=1)
+    >>> [asa,work] = bootstrap_solver(A,num_candidates=1)
     >>> residuals=[]
[…]
