### 2022-02-03 963b1c0
refactor: pull initialization and outer loop from balanced lloyd
