+++ pyamg/graph.py
@@ -150,14 +150,18 @@ def bellman_ford(G, centers, method='standard'):
         raise ValueError('Bellman-Ford is defined only for real weights.')
 
     centers = np.asarray(centers, dtype=np.int32)
+    num_clusters = len(centers)
 
     # allocate space for returns and working arrays
-    distances = np.empty(n, dtype=G.dtype)
-    nearest = np.empty(n, dtype=np.int32)
-    predecessors = np.empty(n, dtype=np.int32)
+    distances = np.full(n, np.inf, dtype=G.dtype) # distance to cluster center (inf)
+    nearest = np.full(n, -1, dtype=np.int32)      # nearest center or cluster membership or index (-1)
+    predecessors = np.full(n, -1, dtype=np.int32) # predecessor on the shortest path (-1)
+    distances[centers] = 0                        # distance = 0 at centers
+    nearest[centers] = np.arange(num_clusters)    # number the membership
+
     if method == 'balanced':
-        predecessors_count = np.empty(n, dtype=np.int32)
-        cluster_size = np.empty(len(centers), dtype=np.int32)
+        predecessors_count = np.full(n, 0, dtype=np.int32)
+        cluster_size = np.full(num_clusters, 1, dtype=np.int32)
 
     if method == 'standard':
         amg_core.bellman_ford(n, G.indptr, G.indices, G.data, centers,  # IN
@@ -166,8 +170,7 @@ def bellman_ford(G, centers, method='standard'):
     elif method == 'balanced':
         amg_core.bellman_ford_balanced(n, G.indptr, G.indices, G.data, centers,  # IN
                                        distances, nearest, predecessors,         # OUT
-                                       predecessors_count, cluster_size,         # OUT
-                                       True)
+                                       predecessors_count, cluster_size)         # OUT
     else:
         raise ValueError(f'Method {method} is not supported in Bellman-Ford')
 
@@ -267,9 +270,11 @@ def balanced_lloyd_cluster(G, centers, maxiter=5, rebalance_iters=5):
 
     Notes
     -----
-    If G has complex values, abs(G) is used instead.
-
-    Only positive edge weights may be used
+    - If G has complex values, abs(G) is used instead.
+    - Only positive edge weights may be used
+    - This version computes improved cluster centers with Floyd-Warshall and
+      also uses a balanced version of Bellman-Ford to try and find
+      nearly-equal-sized clusters.
     """
     G = asgraph(G)
     n = G.shape[0]
@@ -303,40 +308,57 @@ def balanced_lloyd_cluster(G, centers, maxiter=5, rebalance_iters=5):
     # empty() values are initialized in the kernel
     maxsize = int(8*np.ceil((n / num_clusters)))
 
-    Cptr = np.empty(num_clusters, dtype=np.int32)
-    D = np.empty((maxsize, maxsize), dtype=G.dtype)
-    P = np.empty((maxsize, maxsize), dtype=np.int32)
-    CC = np.empty(n, dtype=np.int32)
-    L = np.empty(n, dtype=np.int32)
+    d = np.full(n, np.inf, dtype=G.dtype)         # distance to cluster center (inf)
+    m = np.full(n, -1, dtype=np.int32)            # cluster membership or index (-1)
+    p = np.full(n, -1, dtype=np.int32)            # predecessor on the shortest path (-1)
+    pc = np.full(n, 0, dtype=np.int32)            # predecessor count (0)
+    s = np.full(num_clusters, 1, dtype=np.int32)  # cluster size (1)
+    d[centers] = 0                                # distance = 0 at centers
+    m[centers] = np.arange(num_clusters)          # number the membership
 
-    q = np.empty(maxsize, dtype=G.dtype)
-    d = np.empty(n, dtype=G.dtype)
-    m = np.empty(n, dtype=np.int32)
[…]
