+++ pyamg/amg_core/__init__.py
@@ -11,7 +11,7 @@ from .graph import (maximal_independent_set_serial, maximal_independent_set_para
                     vertex_coloring_mis, vertex_coloring_jones_plassmann,
                     vertex_coloring_LDF,
                     bellman_ford, bellman_ford_balanced,
-                    lloyd_cluster, lloyd_cluster_balanced,
+                    lloyd_cluster,
                     floyd_warshall, center_nodes, most_interior_nodes,
                     maximal_independent_set_k_parallel,
                     breadth_first_search, connected_components)
@@ -58,7 +58,6 @@ __all__ = [
     'bellman_ford',
     'bellman_ford_balanced',
     'lloyd_cluster',
-    'lloyd_cluster_balanced',
     'floyd_warshall',
     'center_nodes',
     'most_interior_nodes',
