+++ pyamg/amg_core/graph_bind.cpp
@@ -315,8 +315,7 @@ bool _bellman_ford_balanced(
        py::array_t<I> & m,
        py::array_t<I> & p,
       py::array_t<I> & pc,
-       py::array_t<I> & s,
-    const bool initialize
+       py::array_t<I> & s
                             )
 {
     auto py_Ap = Ap.unchecked();
@@ -348,8 +347,7 @@ bool _bellman_ford_balanced(
                        _m, m.shape(0),
                        _p, p.shape(0),
                       _pc, pc.shape(0),
-                       _s, s.shape(0),
-               initialize
+                       _s, s.shape(0)
                                        );
 }
 
@@ -433,81 +431,6 @@ void _lloyd_cluster(
                initialize,
                   maxiter
                                );
-}
-
-template<class I, class T>
-void _lloyd_cluster_balanced(
-        const I num_nodes,
-      py::array_t<I> & Ap,
-      py::array_t<I> & Aj,
-      py::array_t<T> & Ax,
-    py::array_t<I> & Cptr,
-       py::array_t<T> & D,
-       py::array_t<I> & P,
-       py::array_t<I> & C,
-       py::array_t<I> & L,
-       py::array_t<T> & q,
-       py::array_t<I> & c,
-       py::array_t<T> & d,
-       py::array_t<I> & m,
-       py::array_t<I> & p,
-      py::array_t<I> & pc,
-       py::array_t<I> & s,
-    const bool initialize,
-        const int maxiter
-                             )
-{
-    auto py_Ap = Ap.unchecked();
-    auto py_Aj = Aj.unchecked();
-    auto py_Ax = Ax.unchecked();
-    auto py_Cptr = Cptr.mutable_unchecked();
-    auto py_D = D.mutable_unchecked();
-    auto py_P = P.mutable_unchecked();
-    auto py_C = C.mutable_unchecked();
-    auto py_L = L.mutable_unchecked();
-    auto py_q = q.mutable_unchecked();
-    auto py_c = c.mutable_unchecked();
-    auto py_d = d.mutable_unchecked();
-    auto py_m = m.mutable_unchecked();
-    auto py_p = p.mutable_unchecked();
-    auto py_pc = pc.mutable_unchecked();
-    auto py_s = s.mutable_unchecked();
-    const I *_Ap = py_Ap.data();
-    const I *_Aj = py_Aj.data();
-    const T *_Ax = py_Ax.data();
-    I *_Cptr = py_Cptr.mutable_data();
-    T *_D = py_D.mutable_data();
-    I *_P = py_P.mutable_data();
-    I *_C = py_C.mutable_data();
-    I *_L = py_L.mutable_data();
-    T *_q = py_q.mutable_data();
-    I *_c = py_c.mutable_data();
-    T *_d = py_d.mutable_data();
-    I *_m = py_m.mutable_data();
-    I *_p = py_p.mutable_data();
-    I *_pc = py_pc.mutable_data();
-    I *_s = py_s.mutable_data();
-
-    return lloyd_cluster_balanced<I, T>(
-                num_nodes,
-                      _Ap, Ap.shape(0),
-                      _Aj, Aj.shape(0),
-                      _Ax, Ax.shape(0),
-                    _Cptr, Cptr.shape(0),
-                       _D, D.shape(0),
-                       _P, P.shape(0),
-                       _C, C.shape(0),
-                       _L, L.shape(0),
-                       _q, q.shape(0),
-                       _c, c.shape(0),
-                       _d, d.shape(0),
-                       _m, m.shape(0),
-                       _p, p.shape(0),
-                      _pc, pc.shape(0),
[…]
