### 2022-02-03 af7200c
refactor: move initilization/outer loop to python in lloyd
