+++ pyamg/amg_core/graph_bind.cpp
@@ -272,8 +272,7 @@ void _bellman_ford(
        py::array_t<I> & c,
        py::array_t<T> & d,
        py::array_t<I> & m,
-       py::array_t<I> & p,
-    const bool initialize
+       py::array_t<I> & p
                    )
 {
     auto py_Ap = Ap.unchecked();
@@ -299,8 +298,7 @@ void _bellman_ford(
                        _c, c.shape(0),
                        _d, d.shape(0),
                        _m, m.shape(0),
-                       _p, p.shape(0),
-               initialize
+                       _p, p.shape(0)
                               );
 }
 
@@ -388,49 +386,6 @@ bool _most_interior_nodes(
                        _m, m.shape(0),
                        _p, p.shape(0)
                                      );
-}
-
-template<class I, class T>
-void _lloyd_cluster(
-        const I num_nodes,
-      py::array_t<I> & Ap,
-      py::array_t<I> & Aj,
-      py::array_t<T> & Ax,
-       py::array_t<I> & c,
-       py::array_t<T> & d,
-       py::array_t<I> & m,
-       py::array_t<I> & p,
-    const bool initialize,
-        const int maxiter
-                    )
-{
-    auto py_Ap = Ap.unchecked();
-    auto py_Aj = Aj.unchecked();
-    auto py_Ax = Ax.unchecked();
-    auto py_c = c.mutable_unchecked();
-    auto py_d = d.mutable_unchecked();
-    auto py_m = m.mutable_unchecked();
-    auto py_p = p.mutable_unchecked();
-    const I *_Ap = py_Ap.data();
-    const I *_Aj = py_Aj.data();
-    const T *_Ax = py_Ax.data();
-    I *_c = py_c.mutable_data();
-    T *_d = py_d.mutable_data();
-    I *_m = py_m.mutable_data();
-    I *_p = py_p.mutable_data();
-
-    return lloyd_cluster<I, T>(
-                num_nodes,
-                      _Ap, Ap.shape(0),
-                      _Aj, Aj.shape(0),
-                      _Ax, Ax.shape(0),
-                       _c, c.shape(0),
-                       _d, d.shape(0),
-                       _m, m.shape(0),
-                       _p, p.shape(0),
-               initialize,
-                  maxiter
-                               );
 }
 
 template<class I, class T, class R>
@@ -531,7 +486,6 @@ PYBIND11_MODULE(graph, m) {
     bellman_ford
     bellman_ford_balanced
     most_interior_nodes
-    lloyd_cluster
     maximal_independent_set_k_parallel
     breadth_first_search
     connected_components
@@ -733,11 +687,11 @@ Notes
 - N is the cluster size)pbdoc");
 
     m.def("bellman_ford", &_bellman_ford<int, int>,
-        py::arg("num_nodes"), py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("Ax").noconvert(), py::arg("c").noconvert(), py::arg("d").noconvert(), py::arg("m").noconvert(), py::arg("p").noconvert(), py::arg("initialize"));
+        py::arg("num_nodes"), py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("Ax").noconvert(), py::arg("c").noconvert(), py::arg("d").noconvert(), py::arg("m").noconvert(), py::arg("p").noconvert());
     m.def("bellman_ford", &_bellman_ford<int, float>,
[…]

[…]
