### 2022-02-04 #4125 @mayrmt
@jedbrown I don't have a conflict of interest and I'm happy to review this submission. Can you please provide more details on the review process for JOSS?
