### 2022-02-04 #4125 @jedbrown
Certainly. The review will be coordinated through a GitHub issue, with possible issues filed directly to upstream. Your review covers the (short) paper, but also the project documentation, and we expect reviewers to experiment with running the software. https://joss.readthedocs.io/en/latest/review_criteria.html
