### 2022-02-04 1b55196
Adding support for using C/F splittings with Rootnode
