+++ pyamg/aggregation/rootnode.py
@@ -4,7 +4,8 @@
 from warnings import warn
 import numpy as np
 from scipy.sparse import csr_matrix, isspmatrix_csr, isspmatrix_bsr,\
-    SparseEfficiencyWarning
+    csc_matrix, SparseEfficiencyWarning
+from scipy.sparse import eye as speye
 
 from ..multilevel import MultilevelSolver
 from ..relaxation.smoothing import change_smoothers
@@ -21,6 +22,7 @@ from .aggregate import standard_aggregation, naive_aggregation, \
     lloyd_aggregation
 from .tentative import fit_candidates
 from .smooth import energy_prolongation_smoother
+from ..classical import split
 
 
 def rootnode_solver(A, B=None, BH=None,
@@ -357,7 +359,7 @@ def _extend_hierarchy(levels, strength, aggregate, smooth, improve_candidates,
     flag, kwargs = unpack_arg(diagonal_dominance)
     if flag:
         C = eliminate_diag_dom_nodes(A, C, **kwargs)
-
+    
     # Compute the aggregation matrix AggOp (i.e., the nodal coarsening of A).
     # AggOp is a boolean matrix, where the sparsity pattern for the k-th column
     # denotes the fine-grid nodes agglomerated into k-th coarse-grid node.
@@ -371,9 +373,54 @@ def _extend_hierarchy(levels, strength, aggregate, smooth, improve_candidates,
     elif fn == 'predefined':
         AggOp = kwargs['AggOp'].tocsr()
         Cnodes = kwargs['Cnodes']
+    elif fn == 'RS':
+        splitting = split.RS(C, **kwargs)
+    elif fn == 'PMIS':
+        splitting = split.PMIS(C, **kwargs)
+    elif fn == 'PMISc':
+        splitting = split.PMISc(C, **kwargs)
+    elif fn == 'CLJP':
+        splitting = split.CLJP(C, **kwargs)
+    elif fn == 'CLJPc':
+        splitting = split.CLJPc(C, **kwargs)
+    elif fn == 'CR':
+        # TODO: adaptive C/F splitting based on  abs(T Bc - Bf).  This could be
+        # incorporated as a new CR type splitting method, or be implemented around 
+        # line 1120 in energy_prolongation_smoothing, where filter_operator is used.
+        # It may belong better here, at least philosophically.
+        splitting = CR(C, **kwargs)
     else:
         raise ValueError(f'Unrecognized aggregation method: {str(fn)}')
 
+    # If using classical splitting, construct needed Rootnode data structures: Cpt_params
+    classical_CF = False
+    if fn in ['RS', 'PMIS', 'PMISc', 'CLJP', 'CLJPc', 'CR']:
+        # TODO: Not sure how to (and if we want) BSR-supernode support for C/F splittings 
+        if isspmatrix_bsr(A):
+            if A.blocksize[0] > 1:
+                raise ValueError(f'Currently, BSR A and classical CF splittings are not supported') 
+
+        classical_CF = True
+        Cpts = (splitting == 1).nonzero()[0]
+        Fpts = (splitting == 0).nonzero()[0]
+        I_C = speye(A.shape[0], A.shape[1], format='csr')
+        I_F = I_C.copy()
+        I_F.data[Cpts] = 0.0
+        I_F.eliminate_zeros()
+        I_C = I_C - I_F
+        I_C.eliminate_zeros()
+        # construct P_I as in get_Cpt_params
+        indices = Cpts.copy()
+        indptr = np.arange(indices.shape[0]+1)
+        ncoarse = Cpts.shape[0] 
+        P_I = csc_matrix((I_C.data.copy(), indices, indptr),
[…]
