+++ pyamg/krylov/_cgne.py
@@ -4,10 +4,9 @@ import warnings
 from warnings import warn
 import numpy as np
 from scipy import sparse
-from scipy.sparse.linalg.isolve.utils import make_system
-from scipy.sparse.linalg.interface import aslinearoperator
-from pyamg.util.linalg import norm
-
+from scipy.sparse.linalg import aslinearoperator
+from ..util.linalg import norm
+from ..util import make_system
 
 def cgne(A, b, x0=None, tol=1e-5, criteria='rr',
          maxiter=None, M=None,
@@ -60,7 +59,7 @@ def cgne(A, b, x0=None, tol=1e-5, criteria='rr',
 
     Notes
     -----
-    The LinearOperator class is in scipy.sparse.linalg.interface.
+    The LinearOperator class is in scipy.sparse.linalg.
     Use this class if you prefer to define A or M as a mat-vec routine
     as opposed to explicitly constructing the matrix.
 
