+++ pyamg/krylov/_bicgstab.py
@@ -2,9 +2,9 @@
 
 import warnings
 import numpy as np
-from scipy.sparse.linalg.isolve.utils import make_system
 from scipy import sparse
-from pyamg.util.linalg import norm
+from ..util.linalg import norm
+from ..util import make_system
 
 
 def bicgstab(A, b, x0=None, tol=1e-5, criteria='rr',
@@ -54,7 +54,7 @@ def bicgstab(A, b, x0=None, tol=1e-5, criteria='rr',
 
     Notes
     -----
-    The LinearOperator class is in scipy.sparse.linalg.interface.
+    The LinearOperator class is in scipy.sparse.linalg.
     Use this class if you prefer to define A or M as a mat-vec routine
     as opposed to explicitly constructing the matrix.
 
