+++ pyamg/krylov/_fgmres.py
@@ -3,11 +3,11 @@
 import warnings
 from warnings import warn
 import numpy as np
-from scipy.sparse.linalg.isolve.utils import make_system
 from scipy.linalg import get_lapack_funcs
 import scipy as sp
-from pyamg.util.linalg import norm
-from pyamg import amg_core
+from ..util.linalg import norm
+from ..util import make_system
+from .. import amg_core
 
 
 def _mysign(x):
@@ -77,7 +77,7 @@ def fgmres(A, b, x0=None, tol=1e-5,
 
     Notes
     -----
-    The LinearOperator class is in scipy.sparse.linalg.interface.
+    The LinearOperator class is in scipy.sparse.linalg.
     Use this class if you prefer to define A or M as a mat-vec routine
     as opposed to explicitly constructing the matrix.
 
