+++ pyamg/util/__init__.py
@@ -4,7 +4,14 @@ from . import linalg
 from . import utils
 from . import params
 
-__all__ = ['linalg', 'utils', 'params']
+try:
+    # scipy >=1.8
+    from scipy.sparse.linalg._isolve.utils import make_system
+except ImportError:
+    # scipy <1.8
+    from scipy.sparse.linalg.isolve.utils import make_system
+
+__all__ = ['linalg', 'utils', 'params', 'make_system']
 
 __doc__ += """
 linalg.py provides some linear algebra functionality not yet found in scipy.
