+++ pyamg/util/__init__.py
@@ -4,14 +4,9 @@ from . import linalg
 from . import utils
 from . import params
 
-try:
-    # scipy >=1.8
-    from scipy.sparse.linalg._isolve.utils import make_system
-except ImportError:
-    # scipy <1.8
-    from scipy.sparse.linalg.isolve.utils import make_system
+from .utils import make_system, upcast
 
-__all__ = ['linalg', 'utils', 'params', 'make_system']
+__all__ = ['linalg', 'utils', 'params', 'make_system', 'upcast']
 
 __doc__ += """
 linalg.py provides some linear algebra functionality not yet found in scipy.
