### 2022-02-07 fee87cd
build: remove pytest from install_requires
