+++ setup.cfg
@@ -43,6 +43,7 @@ python_requires = >=3.6
 install_requires =
     numpy>=1.7.0
     scipy>=0.12.0
+tests_require =
     pytest>=2
 
 [options.packages.find]
