### 2022-02-07 #4125 @lukeolson
@jedbrown @mayrmt scipy 1.8 was released yesterday, causing an import issue with pyamg.  I rolled out a new release today `4.2.1` -> `4.2.2`.  Just a heads-up in case the [version needs to change here](https://joss.readthedocs.io/en/latest/whedon.html?highlight=version#changing-the-software-version).
