### 2022-02-08 #4142 @whedon
[REVIEW]: PyAMG: Algebraic Multigrid Solvers in Python
**Submitting author:** @lukeolson (<a href="http://orcid.org/0000-0002-5283-6104">Luke Olson</a>)
**Repository:** https://github.com/pyamg/pyamg/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v4.2.3

@mayrmt @mattmartineau, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
