### 2022-02-17 839bcac
test: add test for Krylov defaults
