### 2022-02-17 3facb00
fix: move maxiter=None test before comparison
