### 2022-02-17 bae0ffe
test: add accuracy test for tensor kappa
