+++ pyamg/gallery/tests/test_fem.py
@@ -1,6 +1,7 @@
 """Test fem."""
+import pytest
 import numpy as np
-from pyamg.gallery import fem
+from pyamg.gallery import fem, regular_triangle_mesh
 import scipy.sparse.linalg as sla
 
 import os
@@ -248,3 +249,73 @@ class TestGradGradFEM(np.testing.TestCase):
         A = A.tocsr()
         u = sla.spsolve(A, b)
         np.testing.assert_array_almost_equal(u[id3], np.ones(u[id3].shape))
+
+
+def test_gradgrad_kappa():
+    """
+    This test is generated from:
+    from sympy import pi, symbols, diff, Matrix, sin, cos
+    x, y = symbols("x,y")
+    u = sin(2 * pi * x) * sin(2 * pi * y)
+    K = Matrix([[1 + 2*x**2, x * y], [x * y, 2 + 3 * x**2 * y]])
+    gradu = Matrix([[diff(u, x)], [diff(u, y)]])
+    q = K * gradu
+    f = -(diff(q[0], x) + diff(q[1], y))
+    print(f)
+
+    Output:
+    -6*pi*x**2*sin(2*pi*x)*cos(2*pi*y)
+    - 8*pi**2*x*y*cos(2*pi*x)*cos(2*pi*y)
+    - 10*pi*x*sin(2*pi*y)*cos(2*pi*x)
+    - 2*pi*y*sin(2*pi*x)*cos(2*pi*y)
+    + 4*pi**2*(2*x**2 + 1)*sin(2*pi*x)*sin(2*pi*y)
+    + 4*pi**2*(3*x**2*y + 2)*sin(2*pi*x)*sin(2*pi*y)
+    """
+
+    V, E = regular_triangle_mesh(40, 40)
+    mesh = fem.Mesh(V, E)
+    A, b = fem.gradgradform(mesh)
+
+    def f(x, y):
+        pi = np.pi
+        sin = np.sin
+        cos = np.cos
+        return -6*pi*x**2*sin(2*pi*x)*cos(2*pi*y) \
+            - 8*pi**2*x*y*cos(2*pi*x)*cos(2*pi*y) \
+            - 10*pi*x*sin(2*pi*y)*cos(2*pi*x) \
+            - 2*pi*y*sin(2*pi*x)*cos(2*pi*y) \
+            + 4*pi**2*(2*x**2 + 1)*sin(2*pi*x)*sin(2*pi*y) \
+            + 4*pi**2*(3*x**2*y + 2)*sin(2*pi*x)*sin(2*pi*y)
+
+    def g(x, y):
+        return 0*x + 0*y + 0.0
+
+    def kappa(x, y):
+        return np.array([[1 + 2*x**2, x * y], [x * y, 2 + 3 * x**2 * y]])
+
+    def uexact(x, y):
+        pi = np.pi
+        sin = np.sin
+        cos = np.cos
+        return sin(2 * pi * x) * sin(2 * pi * y)
+
+    tol = 1e-12
+    X, Y = V[:, 0], V[:, 1]
+    id1 = np.where(abs(Y) < tol)[0]
+    id2 = np.where(abs(Y-1) < tol)[0]
+    id3 = np.where(abs(X) < tol)[0]
+    id4 = np.where(abs(X-1) < tol)[0]
+
+    bc = [{'id': id1, 'g': g},
+          {'id': id2, 'g': g},
+          {'id': id3, 'g': g},
+          {'id': id4, 'g': g}]
+    mesh = fem.Mesh(V, E)
+    A, b = fem.gradgradform(mesh, f=f, kappa=kappa)
+    A, b = fem.applybc(A, b, mesh, bc)
+
[…]
