### 2022-02-17 a324398
style: flake8 fem and move to helper function
