### 2022-02-18 ea7eb2d
adding flag to rootnode energy min, allowing for use with classical CF style P, i.e., for use with AIR
