### 2022-03-04 7be0c9d
docs: update sphinx conf
