### 2022-03-04 ec6e45d
docs: add sphinx apidoc rst
