+++ docs/source/pyamg.util.rst
@@ -1,0 +1,45 @@
+pyamg.util package
+==================
+
+Submodules
+----------
+
+pyamg.util.bsr\_utils module
+----------------------------
+
+.. automodule:: pyamg.util.bsr_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.util.linalg module
+------------------------
+
+.. automodule:: pyamg.util.linalg
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.util.params module
+------------------------
+
+.. automodule:: pyamg.util.params
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.util.utils module
+-----------------------
+
+.. automodule:: pyamg.util.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyamg.util
+   :members:
+   :undoc-members:
+   :show-inheritance:
