+++ docs/source/pyamg.relaxation.rst
@@ -1,0 +1,45 @@
+pyamg.relaxation package
+========================
+
+Submodules
+----------
+
+pyamg.relaxation.chebyshev module
+---------------------------------
+
+.. automodule:: pyamg.relaxation.chebyshev
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.relaxation.relaxation module
+----------------------------------
+
+.. automodule:: pyamg.relaxation.relaxation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.relaxation.smoothing module
+---------------------------------
+
+.. automodule:: pyamg.relaxation.smoothing
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.relaxation.utils module
+-----------------------------
+
+.. automodule:: pyamg.relaxation.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyamg.relaxation
+   :members:
+   :undoc-members:
+   :show-inheritance:
