+++ docs/source/pyamg.classical.rst
@@ -1,0 +1,45 @@
+pyamg.classical package
+=======================
+
+Submodules
+----------
+
+pyamg.classical.classical module
+--------------------------------
+
+.. automodule:: pyamg.classical.classical
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.classical.cr module
+-------------------------
+
+.. automodule:: pyamg.classical.cr
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.classical.interpolate module
+----------------------------------
+
+.. automodule:: pyamg.classical.interpolate
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.classical.split module
+----------------------------
+
+.. automodule:: pyamg.classical.split
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyamg.classical
+   :members:
+   :undoc-members:
+   :show-inheritance:
