+++ docs/source/pyamg.gallery.rst
@@ -1,0 +1,85 @@
+pyamg.gallery package
+=====================
+
+Submodules
+----------
+
+pyamg.gallery.demo module
+-------------------------
+
+.. automodule:: pyamg.gallery.demo
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.gallery.diffusion module
+------------------------------
+
+.. automodule:: pyamg.gallery.diffusion
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.gallery.elasticity module
+-------------------------------
+
+.. automodule:: pyamg.gallery.elasticity
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.gallery.example module
+----------------------------
+
+.. automodule:: pyamg.gallery.example
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.gallery.fem module
+------------------------
+
+.. automodule:: pyamg.gallery.fem
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.gallery.laplacian module
+------------------------------
+
+.. automodule:: pyamg.gallery.laplacian
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.gallery.mesh module
+-------------------------
+
+.. automodule:: pyamg.gallery.mesh
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.gallery.random\_sparse module
+-----------------------------------
+
+.. automodule:: pyamg.gallery.random_sparse
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.gallery.stencil module
+----------------------------
+
+.. automodule:: pyamg.gallery.stencil
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyamg.gallery
+   :members:
+   :undoc-members:
+   :show-inheritance:
