+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+pyamg
+=====
+
+.. toctree::
+   :maxdepth: 4
+
+   pyamg
