+++ docs/source/pyamg.amg_core.rst
@@ -1,0 +1,69 @@
+pyamg.amg\_core package
+=======================
+
+Submodules
+----------
+
+pyamg.amg\_core.evolution\_strength module
+------------------------------------------
+
+.. automodule:: pyamg.amg_core.evolution_strength
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.amg\_core.graph module
+----------------------------
+
+.. automodule:: pyamg.amg_core.graph
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.amg\_core.krylov module
+-----------------------------
+
+.. automodule:: pyamg.amg_core.krylov
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.amg\_core.linalg module
+-----------------------------
+
+.. automodule:: pyamg.amg_core.linalg
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.amg\_core.relaxation module
+---------------------------------
+
+.. automodule:: pyamg.amg_core.relaxation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.amg\_core.ruge\_stuben module
+-----------------------------------
+
+.. automodule:: pyamg.amg_core.ruge_stuben
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.amg\_core.smoothed\_aggregation module
+--------------------------------------------
+
+.. automodule:: pyamg.amg_core.smoothed_aggregation
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyamg.amg_core
+   :members:
+   :undoc-members:
+   :show-inheritance:
