+++ docs/source/pyamg.vis.rst
@@ -1,0 +1,29 @@
+pyamg.vis package
+=================
+
+Submodules
+----------
+
+pyamg.vis.vis\_coarse module
+----------------------------
+
+.. automodule:: pyamg.vis.vis_coarse
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+pyamg.vis.vtk\_writer module
+----------------------------
+
+.. automodule:: pyamg.vis.vtk_writer
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: pyamg.vis
+   :members:
+   :undoc-members:
+   :show-inheritance:
