+++ docs/source/pyamg.krylov.rst
@@ -1,0 +1,10 @@
+pyamg.krylov package
+====================
+
+Module contents
+---------------
+
+.. automodule:: pyamg.krylov
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
