### 2022-03-04 c558c7c
docs: fix  .h docstrings
