+++ pyamg/amg_core/ruge_stuben.h
@@ -765,8 +765,8 @@ void rs_standard_interpolation(const I n_nodes,
 /* Helper function for compatible relaxation to perform steps 3.1d - 3.1f
  * in Falgout / Brannick (2010).
  *
- * Input:
- * ------
+ * Input
+ * -----
  * A_rowptr : const {int array}
  *      Row pointer for sparse matrix in CSR format.
  * A_colinds : const {int array}
@@ -786,8 +786,8 @@ void rs_standard_interpolation(const I n_nodes,
  * thetacs : const {float}
  *      Threshold for coarse grid candidates from set measure.
  *
- * Returns:
- * --------
+ * Returns
+ * -------
  * Nothing, updated C/F-splitting and corresponding indices modified in place.
  */
 template<class I, class T>
