+++ pyamg/amg_core/evolution_strength.h
@@ -183,7 +183,7 @@ void apply_distance_filter(const I n_row,
  *      modified in place for output
  *
  * Returns
- * ------
+ * -------
  * Tx[i] modified in place, it holds
  * the minimum nonzero value of block i of S
  *
