### 2022-03-04 d3407a9
docs: fix .h docstring; rebind.
