+++ pyamg/amg_core/ruge_stuben.h
@@ -765,8 +765,8 @@ void rs_standard_interpolation(const I n_nodes,
 /* Helper function for compatible relaxation to perform steps 3.1d - 3.1f
  * in Falgout / Brannick (2010).
  *
- * Input
- * -----
+ * Parameters
+ * ----------
  * A_rowptr : const {int array}
  *      Row pointer for sparse matrix in CSR format.
  * A_colinds : const {int array}
