+++ pyamg/amg_core/smoothed_aggregation.h
@@ -639,8 +639,8 @@ void satisfy_constraints_helper(const I rows_per_block,
  * Sp,Sj : {int array}
  *      BSR indptr and indices members for matrix, S
  *
- * Return
- * ------
+ * Returns
+ * -------
  * BtB[i] = B_i.H*B_i in __column__ major format
  * where B_i is B[colindices,:], colindices = all the nonzero
  * column indices for block row i in S
