### 2022-03-04 65f3938
docs: update docstrings and sphinx setup
