+++ docs/source/index.rst
@@ -13,3 +13,4 @@ Reference
     :maxdepth: 4
 
     ref
+    pyamg
