### 2022-03-04 b5fe0de
docs: update docstings.  remove autogenerated.  rebind
