+++ pyamg/amg_core/graph.h
@@ -19,16 +19,25 @@ inline void coreassert(const bool istrue, const std::string &errormsg){
  *  using a greedy serial algorithm
  *
  *  Parameters
- *      num_rows   - number of rows in A (number of vertices)
- *      Ap[]       - CSR row pointer
- *      Aj[]       - CSR index array
- *      active     - value used for active vertices        (input)
- *       C         - value used to mark non-MIS vertices   (output)
- *       F         - value used to mark MIS vertices       (output)
- *      x[]        - state of each vertex
+ *  ----------
+ *  num_rows : int
+ *    Number of rows in A (number of vertices)
+ *  Ap : array
+ *    CSR row pointer
+ *  Aj : array
+ *    CSR index array
+ *  active : float-like
+ *    Value used for active vertices
+ *   C : float-like
+ *    Value used to mark non-MIS vertices
+ *   F : float-like
+ *    Value used to mark MIS vertices
+ *   x : array, inplace output
+ *    State of each vertex
  *
- *
- *  Returns:
+ *  Returns
+ *  --------
+ *  N : int
  *      The number of nodes in the MIS.
  *
  *  Notes:
@@ -550,6 +559,7 @@ I cluster_center(const I a,
  * graph stored in CSR format.
  *
  *  Parameters
+ *  ----------
  *      num_nodes - (IN)    number of nodes (number of rows in A)
  *      Ap[]      - (IN)    CSR row pointer
  *      Aj[]      - (IN)    CSR index array
@@ -557,7 +567,8 @@ I cluster_center(const I a,
  *      d[]       - (INOUT) distance to nearest center
  *     cm[]       - (INOUT) cluster index for each node
  *
- *  References:
+ *  References
+ *  ----------
  *      http://en.wikipedia.org/wiki/Bellman-Ford_algorithm
  */
 template<class I, class T>
