+++ docs/source/pyamg.rst
@@ -1,76 +1,0 @@
-pyamg package
-=============
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   pyamg.aggregation
-   pyamg.amg_core
-   pyamg.classical
-   pyamg.gallery
-   pyamg.krylov
-   pyamg.relaxation
-   pyamg.util
-   pyamg.vis
-
-Submodules
-----------
-
-pyamg.blackbox module
----------------------
-
-.. automodule:: pyamg.blackbox
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyamg.graph module
-------------------
-
-.. automodule:: pyamg.graph
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyamg.graph\_ref module
------------------------
-
-.. automodule:: pyamg.graph_ref
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyamg.multilevel module
------------------------
-
-.. automodule:: pyamg.multilevel
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyamg.strength module
----------------------
-
-.. automodule:: pyamg.strength
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyamg.version module
---------------------
-
-.. automodule:: pyamg.version
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pyamg
-   :members:
-   :undoc-members:
-   :show-inheritance:
