+++ docs/source/pyamg.krylov.rst
@@ -1,10 +1,0 @@
-pyamg.krylov package
-====================
-
-Module contents
----------------
-
-.. automodule:: pyamg.krylov
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
