### 2022-03-04 ce46716
doc: fix .h docstrings; rebind
