+++ pyamg/amg_core/graph_bind.cpp
@@ -441,30 +441,31 @@ Compute a maximal independent set for a graph stored in CSR format
  Parameters
  ----------
  num_rows : int
-   Number of rows in A (number of vertices)
+     Number of rows in A (number of vertices)
  Ap : array
-   CSR row pointer
+     CSR row pointer
  Aj : array
-   CSR index array
+     CSR index array
  active : float-like
-   Value used for active vertices
-  C : float-like
-   Value used to mark non-MIS vertices
-  F : float-like
-   Value used to mark MIS vertices
-  x : array, inplace output
-   State of each vertex
+     Value used for active vertices
+ C : float-like
+     Value used to mark non-MIS vertices
+ F : float-like
+     Value used to mark MIS vertices
+ x : array, inplace output
+     State of each vertex
 
  Returns
  --------
  N : int
      The number of nodes in the MIS.
 
- Notes:
-     Only the vertices with values with x[i] == active are considered
-     when determining the MIS.  Upon return, all active vertices will
-     be assigned the value C or F depending on whether they are in the
-     MIS or not.)pbdoc");
+ Notes
+ -----
+ Only the vertices with values with x[i] == active are considered
+ when determining the MIS.  Upon return, all active vertices will
+ be assigned the value C or F depending on whether they are in the
+ MIS or not.)pbdoc");
 
     m.def("maximal_independent_set_parallel", &_maximal_independent_set_parallel<int, int, double>,
         py::arg("num_rows"), py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("active"), py::arg("C"), py::arg("F"), py::arg("x").noconvert(), py::arg("y").noconvert(), py::arg("max_iters"),
@@ -473,26 +474,37 @@ Compute a maximal independent set for a graph stored in CSR format
  using a variant of Luby's parallel MIS algorithm
 
  Parameters
-     num_rows   - number of rows in A (number of vertices)
-     Ap[]       - CSR row pointer
-     Aj[]       - CSR index array
-     active     - value used for active vertices        (input)
-      C         - value used to mark non-MIS vertices   (output)
-      F         - value used to mark MIS vertices       (output)
-     x[]        - state of each vertex
-     y[]        - random values for each vertex
-     max_iters  - maximum number of iterations
-                  by default max_iters=-1 and no limit
-                  is imposed
+ ----------
+ num_rows : int
+     number of rows in A (number of vertices)
+ Ap : array
+     CSR row pointer
+ Aj : array
+     CSR index array
+ active : float
+     value used for active vertices
+  C : float
+     value used to mark non-MIS vertices
+  F : float
+     value used to mark MIS vertices
+ x : array, output
+     state of each vertex
+ y : array
+     random values for each vertex
+ max_iters : int
+     maximum number of iterations By default max_iters=-1 and no limit is imposed
 
- Returns:
+ Returns
+ -------
+ N : int
      The number of nodes in the MIS.
 
- Notes:
-     Only the vertices with values with x[i] == active are considered
-     when determining the MIS.  Upon return, all active vertices will
-     be assigned the value C or F depending on whether they are in the
-     MIS or not.)pbdoc");
+ Notes
+ -----
+ Only the vertices with values with x[i] == active are considered
+ when determining the MIS.  Upon return, all active vertices will
+ be assigned the value C or F depending on whether they are in the
+ MIS or not.)pbdoc");
 
     m.def("vertex_coloring_mis", &_vertex_coloring_mis<int, int>,
         py::arg("num_rows"), py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("x").noconvert(),
