### 2022-03-04 289e1b7
docs: update .h docstrings in relaxation; rebind
