### 2022-03-05 a0f5108
docs: update .h docstrings in graph.h; rebind
