+++ pyamg/amg_core/graph.h
@@ -356,39 +356,44 @@ T vertex_coloring_LDF(const I num_rows,
 /*
  * Compute the incidence matrix for a clustering
  *
- *      I = Incidence matrix between nodes and clusters (num_nodes x num_clusters)
+ * Parameters
+ * ----------
+ * num_nodes : int
+ *     number of nodes
+ * num_clusters : int
+ *     number of clusters
+ * cm : array, num_nodes
+ *     cluster index for each node
+ * ICp : arrayt, num_clusters+1, inplace
+ *     CSC column pointer array for I
+ * ICi : array, num_nodes, inplace
+ *     CSC column indexes for I
+ * L : array, num_nodes, inplace
+ *     Local index mapping
+ *
+ * Notes
+ * -----
+ * I = Incidence matrix between nodes and clusters (num_nodes x num_clusters)
  * I[i,a] = 1 if node i is in cluster a, otherwise 0
  *
- *     Cluster indexes: a,b,c in 1..num_clusters
+ * Cluster indexes: a,b,c in 1..num_clusters
  * Global node indexes: i,j,k in 1..num_rows
- *  Local node indexes: pair (a,m) where a is cluster and m in 1..num_nodes_in_cluster
+ * Local node indexes: pair (a,m) where a is cluster and m in 1..num_nodes_in_cluster
  *
  * We store I in both CSC and CSR formats because we want to be able
  * to map global <-> local node indexes. However, I in CSR format is
  * simply the cm array, so we only need to compute CSC format.
  *
- * IC = (ICp,ICi)    = I in CSC format (don't store ICx because it's
- *                     always 1).
+ * IC = (ICp,ICi)    = I in CSC format (don't store ICx because it's always 1).
  *
  * IR = (IRa) = (cm) = I in CSR format (don't store IRp because we
- *                     have exactly one nonzero entry per row, and
- *                     don't store IRx because it's always 1). This is
- *                     just the cm array.
+ * have exactly one nonzero entry per row, and don't store IRx because it's always 1). This is
+ * just the cm array.
  *
  * Converting local (a,m) -> global i:   i = ICi[ICp[a] + m]
- *
- * Converting global i -> local (a,m):   a = cm[i]
- *                                       m = L[i]
+ * Converting global i -> local (a,m):   a = cm[i], m = L[i]
  *
  * L is an additional vector (length num_rows) to store local indexes.
- *
- *  Parameters
- *      num_nodes         - (IN)  number of nodes
- *      num_clusters      - (IN)  number of clusters
- *     cm[num_nodes]      - (IN)  cluster index for each node
- *    ICp[num_clusters+1] - (OUT) CSC column pointer array for I
- *    ICi[num_nodes]      - (OUT) CSC column indexes for I
- *      L[num_nodes]      - (OUT) Local index mapping
  */
 template<class I>
 void cluster_node_incidence(const I num_nodes,
@@ -471,22 +476,36 @@ void cluster_node_incidence(const I num_nodes,
  * Apply Floyd–Warshall to cluster "a" and use the result to find the
  * cluster center
  *
- *  Parameters
- *      a                  - (IN) cluster index to find the center of
- *      num_nodes          - (IN) number of nodes
- *      num_clusters       - (IN) number of clusters
- *      Ap[]               - (IN) CSR row pointer
- *      Aj[]               - (IN) CSR index array
- *      Ax[]               - (IN) CSR data array (edge lengths)
- *      cm[num_nodes]      - (IN) cluster index for each node
- *     ICp[num_clusters+1] - (IN) CSC column pointer array for I
- *     ICi[num_nodes]      - (IN) CSC column indexes for I
- *       L[num_nodes]      - (IN) Local index mapping
[…]
