### 2022-03-05 3cdc7bf
docs: update docstrings in graph.h; rebind
