+++ pyamg/amg_core/graph.h
@@ -504,11 +504,11 @@ void cluster_node_incidence(const I num_nodes,
  * i : int
  *     global node index of center of cluster a
  *
- *  References
- *  ----------
- *      https://en.wikipedia.org/wiki/Graph_center
- *      https://en.wikipedia.org/wiki/Floyd–Warshall_algorithm
- *      https://en.wikipedia.org/wiki/Distance_(graph_theory)
+ * References
+ * ----------
+ * .. [1] Graph Center:   https://en.wikipedia.org/wiki/Graph_center
+ * .. [2] Floyd-Warshall: https://en.wikipedia.org/wiki/Floyd–Warshall_algorithm
+ * .. [3] Graph Distance: https://en.wikipedia.org/wiki/Distance_(graph_theory)
  */
 template<class I, class T>
 I cluster_center(const I a,
@@ -589,18 +589,25 @@ I cluster_center(const I a,
  * Apply one iteration of Bellman-Ford iteration on a distance
  * graph stored in CSR format.
  *
- *  Parameters
- *  ----------
- *      num_nodes - (IN)    number of nodes (number of rows in A)
- *      Ap[]      - (IN)    CSR row pointer
- *      Aj[]      - (IN)    CSR index array
- *      Ax[]      - (IN)    CSR data array (edge lengths)
- *      d[]       - (INOUT) distance to nearest center
- *     cm[]       - (INOUT) cluster index for each node
+ * Parameters
+ * ----------
+ * num_nodes : int
+ *     number of nodes (number of rows in A)
+ * Ap : array
+ *     CSR row pointer
+ * Aj : array
+ *     CSR index array
+ * Ax : array
+ *     CSR data array (edge lengths)
+ * d : array, inplace
+ *     distance to nearest center
+ * cm : array, inplace
+ *     cluster index for each node
  *
- *  References
- *  ----------
- *      http://en.wikipedia.org/wiki/Bellman-Ford_algorithm
+ * References
+ * ----------
+ * .. Bellman-Ford Wikipedia:
+ *     http://en.wikipedia.org/wiki/Bellman-Ford_algorithm
  */
 template<class I, class T>
 void bellman_ford(const I num_nodes,
