### 2022-03-05 5f8f6d8
docs: update docstrings in graph.h and sa; rebind
