+++ pyamg/amg_core/smoothed_aggregation_bind.cpp
@@ -361,70 +361,95 @@ PYBIND11_MODULE(smoothed_aggregation, m) {
         py::arg("n_row"), py::arg("theta"), py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("Ax").noconvert(), py::arg("Sp").noconvert(), py::arg("Sj").noconvert(), py::arg("Sx").noconvert(),
 R"pbdoc(
 Compute a strength of connection matrix using the standard symmetric
- Smoothed Aggregation heuristic.  Both the input and output matrices
- are stored in CSR format.  A nonzero connection A[i,j] is considered
- strong if:
+Smoothed Aggregation heuristic.  Both the input and output matrices
+are stored in CSR format.  A nonzero connection A[i,j] is considered
+strong if:
 
-     abs(A[i,j]) >= theta * sqrt( abs(A[i,i]) * abs(A[j,j]) )
+..
+    abs(A[i,j]) >= theta * sqrt( abs(A[i,i]) * abs(A[j,j]) )
 
- The strength of connection matrix S is simply the set of nonzero entries
- of A that qualify as strong connections.
+The strength of connection matrix S is simply the set of nonzero entries
+of A that qualify as strong connections.
 
- Parameters
-     num_rows   - number of rows in A
-     theta      - stength of connection tolerance
-     Ap[]       - CSR row pointer
-     Aj[]       - CSR index array
-     Ax[]       - CSR data array
-     Sp[]       - (output) CSR row pointer
-     Sj[]       - (output) CSR index array
-     Sx[]       - (output) CSR data array
+Parameters
+----------
+num_rows : int
+    number of rows in A
+theta : float
+    stength of connection tolerance
+Ap : array
+    CSR row pointer
+Aj : array
+    CSR index array
+Ax : array
+    CSR data array
+Sp : array, inplace
+    CSR row pointer
+Sj : array, inplace
+    CSR index array
+Sx : array, inplace
+    CSR data array
 
-
- Returns:
-     Nothing, S will be stored in Sp, Sj, Sx
-
- Notes:
-     Storage for S must be preallocated.  Since S will consist of a subset
-     of A's nonzero values, a conservative bound is to allocate the same
-     storage for S as is used by A.)pbdoc");
+Notes
+-----
+Storage for S must be preallocated.  Since S will consist of a subset
+of A's nonzero values, a conservative bound is to allocate the same
+storage for S as is used by A.)pbdoc");
 
     m.def("standard_aggregation", &_standard_aggregation<int>,
         py::arg("n_row"), py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("x").noconvert(), py::arg("y").noconvert(),
 R"pbdoc(
 Compute aggregates for a matrix A stored in CSR format
 
-Parameters:
-  n_row         - number of rows in A
-  Ap[n_row + 1] - CSR row pointer
-  Aj[nnz]       - CSR column indices
-   x[n_row]     - aggregate numbers for each node
-   y[n_row]     - will hold Cpts upon return
+Parameters
+----------
+n_row : int
+    number of rows in A
+Ap : array, n_row + 1
+    CSR row pointer
+Aj : array, nnz
+    CSR column indices
+x : array, n_row, inplace
+    aggregate numbers for each node
+y : array, n_row, inplace
+    will hold Cpts upon return
 
-Returns:
[…]

[…]
