### 2022-03-05 d24288c
docs: update docstrings in graph.h; rebind
