+++ pyamg/amg_core/graph.h
@@ -91,9 +91,9 @@ I maximal_independent_set_serial(const I num_rows,
  *      CSR index array
  *  active : float
  *      value used for active vertices
- *   C : float
+ *  C : float
  *      value used to mark non-MIS vertices
- *   F : float
+ *  F : float
  *      value used to mark MIS vertices
  *  x : array, output
  *      state of each vertex
@@ -248,22 +248,29 @@ void vertex_coloring_first_fit(const I num_rows,
 /*
  * Compute a vertex coloring of a graph using the Jones-Plassmann algorithm
  *
- *  Parameters
- *      num_rows   - number of rows in A (number of vertices)
- *      Ap[]       - CSR row pointer
- *      Aj[]       - CSR index array
- *      x[]        - color of each vertex
- *      y[]        - initial random values for each vertex
+ * Parameters
+ * ----------
+ * num_rows : int
+ *     number of rows in A (number of vertices)
+ * Ap : array
+ *     CSR row pointer
+ * Aj : array
+ *     CSR index array
+ * x : array, inplace
+ *     color of each vertex
+ * y : array
+ *     initial random values for each vertex
  *
- *  Notes:
- *      Arrays x and y will be overwritten
+ * Notes
+ * -----
+ *     Arrays x and y will be overwritten
  *
- *  References:
- *      Mark T. Jones and Paul E. Plassmann
- *      A Parallel Graph Coloring Heuristic
- *      SIAM Journal on Scientific Computing 14:3 (1993) 654--669
- *      http://citeseer.ist.psu.edu/jones92parallel.html
- *
+ * References
+ * ----------
+ * .. [Jones92] Mark T. Jones and Paul E. Plassmann
+ *    A Parallel Graph Coloring Heuristic
+ *    SIAM Journal on Scientific Computing 14:3 (1993) 654--669
+ *    http://citeseer.ist.psu.edu/jones92parallel.html
  */
 template<class I, class T, class R>
 T vertex_coloring_jones_plassmann(const I num_rows,
@@ -299,18 +306,20 @@ T vertex_coloring_jones_plassmann(const I num_rows,
  * Compute a vertex coloring of a graph using the parallel
  * Largest-Degree-First (LDF) algorithm
  *
- *  Parameters
- *      num_rows   - number of rows in A (number of vertices)
- *      Ap[]       - CSR row pointer
- *      Aj[]       - CSR index array
- *      x[]        - color of each vertex
- *      y[]        - initial random values for each vertex
+ * Parameters
+ * ----------
+ *     num_rows   - number of rows in A (number of vertices)
+ *     Ap[]       - CSR row pointer
+ *     Aj[]       - CSR index array
+ *     x[]        - color of each vertex
+ *     y[]        - initial random values for each vertex
  *
- *   References:
- *     J. R. Allwright and R. Bordawekar and P. D. Coddington and K. Dincer and C. L. Martin
- *     A Comparison of Parallel Graph Coloring Algorithms
- *     DRAFT SCCS-666
- *     http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.45.4650
+ * References
+ * ----------
+ * .. [LDF] J. R. Allwright and R. Bordawekar and P. D. Coddington and K. Dincer and C. L. Martin
+ *    A Comparison of Parallel Graph Coloring Algorithms
+ *    DRAFT SCCS-666
+ *    http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.45.4650
  *
  */
 template<class I, class T, class R>
@@ -606,8 +615,7 @@ I cluster_center(const I a,
  *
  * References
[…]
