+++ pyamg/amg_core/graph_bind.cpp
@@ -483,9 +483,9 @@ Compute a maximal independent set for a graph stored in CSR format
      CSR index array
  active : float
      value used for active vertices
-  C : float
+ C : float
      value used to mark non-MIS vertices
-  F : float
+ F : float
      value used to mark MIS vertices
  x : array, output
      state of each vertex
@@ -525,21 +525,29 @@ Compute a vertex coloring for a graph stored in CSR format.
 R"pbdoc(
 Compute a vertex coloring of a graph using the Jones-Plassmann algorithm
 
- Parameters
-     num_rows   - number of rows in A (number of vertices)
-     Ap[]       - CSR row pointer
-     Aj[]       - CSR index array
-     x[]        - color of each vertex
-     y[]        - initial random values for each vertex
+Parameters
+----------
+num_rows : int
+    number of rows in A (number of vertices)
+Ap : array
+    CSR row pointer
+Aj : array
+    CSR index array
+x : array, inplace
+    color of each vertex
+y : array
+    initial random values for each vertex
 
- Notes:
-     Arrays x and y will be overwritten
+Notes
+-----
+    Arrays x and y will be overwritten
 
- References:
-     Mark T. Jones and Paul E. Plassmann
-     A Parallel Graph Coloring Heuristic
-     SIAM Journal on Scientific Computing 14:3 (1993) 654--669
-     http://citeseer.ist.psu.edu/jones92parallel.html)pbdoc");
+References
+----------
+.. [Jones92] Mark T. Jones and Paul E. Plassmann
+   A Parallel Graph Coloring Heuristic
+   SIAM Journal on Scientific Computing 14:3 (1993) 654--669
+   http://citeseer.ist.psu.edu/jones92parallel.html)pbdoc");
 
     m.def("vertex_coloring_LDF", &_vertex_coloring_LDF<int, int, double>,
         py::arg("num_rows"), py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("x").noconvert(), py::arg("y").noconvert(),
@@ -547,18 +555,20 @@ R"pbdoc(
 Compute a vertex coloring of a graph using the parallel
 Largest-Degree-First (LDF) algorithm
 
- Parameters
-     num_rows   - number of rows in A (number of vertices)
-     Ap[]       - CSR row pointer
-     Aj[]       - CSR index array
-     x[]        - color of each vertex
-     y[]        - initial random values for each vertex
+Parameters
+----------
+    num_rows   - number of rows in A (number of vertices)
+    Ap[]       - CSR row pointer
+    Aj[]       - CSR index array
+    x[]        - color of each vertex
+    y[]        - initial random values for each vertex
 
-  References:
-    J. R. Allwright and R. Bordawekar and P. D. Coddington and K. Dincer and C. L. Martin
-    A Comparison of Parallel Graph Coloring Algorithms
-    DRAFT SCCS-666
-    http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.45.4650)pbdoc");
+References
+----------
+.. [LDF] J. R. Allwright and R. Bordawekar and P. D. Coddington and K. Dincer and C. L. Martin
+   A Comparison of Parallel Graph Coloring Algorithms
+   DRAFT SCCS-666
+   http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.45.4650)pbdoc");
 
     m.def("cluster_node_incidence", &_cluster_node_incidence<int>,
         py::arg("num_nodes"), py::arg("num_clusters"), py::arg("cm").noconvert(), py::arg("ICp").noconvert(), py::arg("ICi").noconvert(), py::arg("L").noconvert(),
@@ -675,8 +685,7 @@ cm : array, inplace
 
 References
 ----------
-.. Bellman-Ford Wikipedia:
[…]
