### 2022-03-05 fbea996
docs: update docstrings in relaxation.h; rebind
