+++ pyamg/amg_core/relaxation_bind.cpp
@@ -555,29 +555,41 @@ PYBIND11_MODULE(relaxation, m) {
         py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("Ax").noconvert(), py::arg("x").noconvert(), py::arg("b").noconvert(), py::arg("row_start"), py::arg("row_stop"), py::arg("row_step"),
 R"pbdoc(
 Perform one iteration of Gauss-Seidel relaxation on the linear
- system Ax = b, where A is stored in CSR format and x and b
- are column vectors.
+system Ax = b, where A is stored in CSR format and x and b
+are column vectors.
 
- The unknowns are swept through according to the slice defined
- by row_start, row_end, and row_step.  These options are used
- to implement standard forward and backward sweeps, or sweeping
- only a subset of the unknowns.  A forward sweep is implemented
- with gauss_seidel(Ap, Aj, Ax, x, b, 0, N, 1) where N is the
- number of rows in matrix A.  Similarly, a backward sweep is
- implemented with gauss_seidel(Ap, Aj, Ax, x, b, N, -1, -1).
+Parameters
+----------
+Ap : array
+    CSR row pointer
+Aj : array
+    CSR index array
+Ax : array
+    CSR data array
+x : array, inplace
+    approximate solution
+b : array
+    right hand side
+row_start : int
+    beginning of the sweep
+row_stop : int
+    end of the sweep (i.e. one past the last unknown)
+row_step : int
+    stride used during the sweep (may be negative)
 
- Parameters
-     Ap[]       - CSR row pointer
-     Aj[]       - CSR index array
-     Ax[]       - CSR data array
-     x[]        - approximate solution
-     b[]        - right hand side
-     row_start  - beginning of the sweep
-     row_stop   - end of the sweep (i.e. one past the last unknown)
-     row_step   - stride used during the sweep (may be negative)
+Returns
+-------
+Nothing, x will be modified inplace
 
- Returns:
-     Nothing, x will be modified in place)pbdoc");
+Notes
+-----
+The unknowns are swept through according to the slice defined
+by row_start, row_end, and row_step.  These options are used
+to implement standard forward and backward sweeps, or sweeping
+only a subset of the unknowns.  A forward sweep is implemented
+with gauss_seidel(Ap, Aj, Ax, x, b, 0, N, 1) where N is the
+number of rows in matrix A.  Similarly, a backward sweep is
+implemented with gauss_seidel(Ap, Aj, Ax, x, b, N, -1, -1).)pbdoc");
 
     m.def("bsr_gauss_seidel", &_bsr_gauss_seidel<int, float, float>,
         py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("Ax").noconvert(), py::arg("x").noconvert(), py::arg("b").noconvert(), py::arg("row_start"), py::arg("row_stop"), py::arg("row_step"), py::arg("blocksize"));
@@ -589,26 +601,37 @@ Perform one iteration of Gauss-Seidel relaxation on the linear
         py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("Ax").noconvert(), py::arg("x").noconvert(), py::arg("b").noconvert(), py::arg("row_start"), py::arg("row_stop"), py::arg("row_step"), py::arg("blocksize"),
 R"pbdoc(
 Perform one iteration of Gauss-Seidel relaxation on the linear
- system Ax = b, where A is stored in Block CSR format and x and b
- are column vectors.  This method applies point-wise relaxation
- to the BSR as opposed to \"block relaxation\".
+system Ax = b, where A is stored in Block CSR format and x and b
+are column vectors.  This method applies point-wise relaxation
+to the BSR as opposed to \"block relaxation\".
 
- Refer to gauss_seidel for additional information regarding
[…]
