+++ pyamg/amg_core/linalg.h
@@ -132,16 +132,17 @@ inline std::complex<double> zero_imag(std::complex<double>& x)
  *
  * Parameters
  * ----------
- * x : {float|complex array}
- *      n-vector
- * y : {float|complex array}
- *      n-vector
- * n : {int}
- *      size of x and y
+ * x : array
+ *     n-vector
+ * y : array
+ *     n-vector
+ * n : int
+ *     size of x and y
  *
- * Return
- * ------
- * conjugate(x).T y
+ * Returns
+ * -------
+ * float
+ *     conjugate(x).T y
  *
  */
 template<class I, class T>
@@ -158,16 +159,17 @@ inline T dot_prod(const T x[], const T y[], const I n)
  *
  * Parameters
  * ----------
- * x : {float|complex array}
- *      n-vector
- * n : {int}
- *      size of x and y
- * normx : {scalar}
- *      output value
+ * x : array
+ *     n-vector
+ * n : int
+ *     size of x and y
+ * normx : float
+ *     output value
  *
- * Return
- * ------
- * normx = sqrt( <x, x> )
+ * Returns
+ * -------
+ * float
+ *     normx = sqrt( <x, x> )
  *
  */
 template<class I, class T, class F>
@@ -181,18 +183,19 @@ inline void norm(const T x[], const I n, F &normx)
  *
  * Parameters
  * ----------
- * x : {float|complex array}
- *      n-vector
- * y : {float|complex array}
- *      n-vector
- * n : {int}
- *      size of x and y
- * alpha : {scalar}
+ * x : array
+ *     n-vector
+ * y : array
+ *     n-vector
+ * n : int
+ *     size of x and y
+ * alpha : float
+ *     value to scale with
  *
- * Return
- * ------
- * x = x + alpha*y
- *
+ * Returns
+ * -------
+ * x : float
+ *     x = x + alpha*y
  */
 template<class I, class T>
 inline void axpy(T x[], const T y[], const T alpha, const I n)
@@ -206,15 +209,17 @@ inline void axpy(T x[], const T y[], const T alpha, const I n)
  *
  * Parameters
  * ----------
- * Ax : {float|complex array}
+ * Ax : array
  *      m x n dense array
- * Bx : {float|complex array}
+ * Bx :array
  *      m x n dense array
- * m,n : {int}
- *      Dimensions of Ax and Bx
+ * m : int
+ *      Dimensions of Ax
+ * n : int
+ *      Dimensions of Bx
  *
- * Return
- * ------
+ * Returns
+ * -------
  * Bx is overwritten with the transpose of Ax
  *
  * Notes
@@ -309,52 +314,49 @@ inline void transpose(const T Ax[], T Bx[], const I m, const I n)
  *
  * Parameters
  * ----------
- * Ax : {float|complex array}
+ * Ax : array
  *      Stored in row major
- * Arows : {int}
+ * Arows : int
  *      Number of rows of A
- * Acols : {int}
+ * Acols : int
  *      Number of columns of A
- * Atrans : {char}
+ * Atrans : char
  *      Not Used
- * Bx : {float|complex array}
+ * Bx : array
  *      Stored in col major
- * Brows : {int}
+ * Brows : int
  *      Number of rows of B
- * Bcols : {int}
+ * Bcols : int
  *      Number of columns of B
- * Btrans : {char}
+ * Btrans : char
  *      Supported, essentially Btrans='F' assumes
  *      B is in column major, and Brans='T' assumes
  *      B is in row major
- * Sx : {float|complex array}
+ * Sx : array
  *      Output array, Contents are overwritten
- * Srows : {int}
+ * Srows : int
  *      Number of rows of S
- * Scols : {int}
+ * Scols : int
  *      Number of columns of S
- * Strans : {char}
+ * Strans : char
[…]

[…]
