### 2022-03-05 b4bc4b3
docs: update docstrings in ruge_stuben.h; rebind
