+++ pyamg/amg_core/ruge_stuben_bind.cpp
@@ -351,28 +351,39 @@ Compute a strength of connection matrix using the classical strength
  matrices are stored in CSR format.  An off-diagonal nonzero entry
  A[i,j] is considered strong if:
 
+ ..
      |A[i,j]| >= theta * max( |A[i,k]| )   where k != i
 
 Otherwise, the connection is weak.
 
- Parameters
-     num_rows   - number of rows in A
-     theta      - stength of connection tolerance
-     Ap[]       - CSR row pointer
-     Aj[]       - CSR index array
-     Ax[]       - CSR data array
-     Sp[]       - (output) CSR row pointer
-     Sj[]       - (output) CSR index array
-     Sx[]       - (output) CSR data array
+Parameters
+----------
+num_rows : int
+    number of rows in A
+theta : float
+    stength of connection tolerance
+Ap : array
+    CSR row pointer
+Aj : array
+    CSR index array
+Ax : array
+    CSR data array
+Sp : array
+    CSR row pointer
+Sj : array
+    CSR index array
+Sx : array
+    CSR data array
 
+Returns
+-------
+Nothing, S will be stored in Sp, Sj, Sx
 
- Returns:
-     Nothing, S will be stored in Sp, Sj, Sx
-
- Notes:
-     Storage for S must be preallocated.  Since S will consist of a subset
-     of A's nonzero values, a conservative bound is to allocate the same
-     storage for S as is used by A.)pbdoc");
+Notes
+-----
+Storage for S must be preallocated.  Since S will consist of a subset
+of A's nonzero values, a conservative bound is to allocate the same
+storage for S as is used by A.)pbdoc");
 
     m.def("classical_strength_of_connection_min", &_classical_strength_of_connection_min<int, float>,
         py::arg("n_row"), py::arg("theta"), py::arg("Ap").noconvert(), py::arg("Aj").noconvert(), py::arg("Ax").noconvert(), py::arg("Sp").noconvert(), py::arg("Sj").noconvert(), py::arg("Sx").noconvert());
@@ -392,15 +403,22 @@ R"pbdoc(
 R"pbdoc(
 Compute the maximum in magnitude row value for a CSR matrix
 
- Parameters
-     num_rows   - number of rows in A
-     Ap[]       - CSR row pointer
-     Aj[]       - CSR index array
-     Ax[]       - CSR data array
-      x[]       - num_rows array
+Parameters
+----------
+num_rows : int
+    number of rows in A
+x : array, inplace
+     num_rows array
+Ap : array
+    CSR row pointer
+Aj : array
+    CSR index array
+Ax : array
+    CSR data array
 
- Returns:
-     Nothing, x[i] will hold row i's maximum magnitude entry)pbdoc");
+Returns
+-------
+Nothing, x[i] will hold row i's maximum magnitude entry)pbdoc");
 
     m.def("rs_cf_splitting", &_rs_cf_splitting<int>,
         py::arg("n_nodes"), py::arg("C_rowptr").noconvert(), py::arg("C_colinds").noconvert(), py::arg("Tp").noconvert(), py::arg("Tj").noconvert(), py::arg("influence").noconvert(), py::arg("splitting").noconvert(),
@@ -411,18 +429,27 @@ and its transpose T, are stored in CSR format.  Upon return, the  splitting
 array will consist of zeros and ones, where C-nodes (coarse nodes) are
 marked with the value 1 and F-nodes (fine nodes) with the value 0.
 
-Parameters:
-  n_nodes   - number of rows in A
-  C_rowptr[]      - CSR row pointer array for SOC matrix
-  C_colinds[]      - CSR column index array for SOC matrix
-  Tp[]      - CSR row pointer array for transpose of SOC matrix
-  Tj[]      - CSR column index array for transpose of SOC matrix
[…]
