+++ pyamg/amg_core/ruge_stuben.h
@@ -23,29 +23,39 @@
  *  matrices are stored in CSR format.  An off-diagonal nonzero entry
  *  A[i,j] is considered strong if:
  *
+ *  ..
  *      |A[i,j]| >= theta * max( |A[i,k]| )   where k != i
  *
  * Otherwise, the connection is weak.
  *
- *  Parameters
- *      num_rows   - number of rows in A
- *      theta      - stength of connection tolerance
- *      Ap[]       - CSR row pointer
- *      Aj[]       - CSR index array
- *      Ax[]       - CSR data array
- *      Sp[]       - (output) CSR row pointer
- *      Sj[]       - (output) CSR index array
- *      Sx[]       - (output) CSR data array
+ * Parameters
+ * ----------
+ * num_rows : int
+ *     number of rows in A
+ * theta : float
+ *     stength of connection tolerance
+ * Ap : array
+ *     CSR row pointer
+ * Aj : array
+ *     CSR index array
+ * Ax : array
+ *     CSR data array
+ * Sp : array
+ *     CSR row pointer
+ * Sj : array
+ *     CSR index array
+ * Sx : array
+ *     CSR data array
  *
+ * Returns
+ * -------
+ * Nothing, S will be stored in Sp, Sj, Sx
  *
- *  Returns:
- *      Nothing, S will be stored in Sp, Sj, Sx
- *
- *  Notes:
- *      Storage for S must be preallocated.  Since S will consist of a subset
- *      of A's nonzero values, a conservative bound is to allocate the same
- *      storage for S as is used by A.
- *
+ * Notes
+ * -----
+ * Storage for S must be preallocated.  Since S will consist of a subset
+ * of A's nonzero values, a conservative bound is to allocate the same
+ * storage for S as is used by A.
  */
 template<class I, class T, class F>
 void classical_strength_of_connection_abs(const I n_row,
@@ -148,17 +158,24 @@ void classical_strength_of_connection_min(const I n_row,
 }
 
 /*
- *  Compute the maximum in magnitude row value for a CSR matrix
+ * Compute the maximum in magnitude row value for a CSR matrix
  *
- *  Parameters
- *      num_rows   - number of rows in A
- *      Ap[]       - CSR row pointer
- *      Aj[]       - CSR index array
- *      Ax[]       - CSR data array
- *       x[]       - num_rows array
+ * Parameters
+ * ----------
+ * num_rows : int
+ *     number of rows in A
+ * x : array, inplace
+ *      num_rows array
+ * Ap : array
+ *     CSR row pointer
+ * Aj : array
+ *     CSR index array
+ * Ax : array
+ *     CSR data array
  *
- *  Returns:
- *      Nothing, x[i] will hold row i's maximum magnitude entry
+ * Returns
+ * -------
+ * Nothing, x[i] will hold row i's maximum magnitude entry
  *
  */
 template<class I, class T, class F>
@@ -191,19 +208,27 @@ void maximum_row_value(const I n_row,
  * array will consist of zeros and ones, where C-nodes (coarse nodes) are
  * marked with the value 1 and F-nodes (fine nodes) with the value 0.
  *
- * Parameters:
- *   n_nodes   - number of rows in A
- *   C_rowptr[]      - CSR row pointer array for SOC matrix
- *   C_colinds[]      - CSR column index array for SOC matrix
- *   Tp[]      - CSR row pointer array for transpose of SOC matrix
- *   Tj[]      - CSR column index array for transpose of SOC matrix
- *   influence - array that influences splitting (values stored here are
- *               added to lambda for each point)
- *   splitting - array to store the C/F splitting
+ * Parameters
+ * ----------
+ * n_nodes : int
[…]
