### 2022-03-05 3c4dc7e
docs: add config for rtds
