### 2022-03-05 8d239a5
docs: change sphinx dep from m2r to m2r2
