### 2022-03-08 51ba714
test: add test for scipy bellman_ford
Also note the "transpose"
