+++ pyamg/amg_core/linalg.h
@@ -128,6 +128,26 @@ inline std::complex<double> zero_imag(std::complex<double>& x)
  *      templated for pyamg's complex class, float and double
  *******************************************************************/
 
+/*
+ * Return row-major index from 2d array index, A[row,col].
+ */
+template<class I>
+inline I row_major(const I row, const I col, const I num_cols)
+{
+    return row*num_cols + col;
+}
+
+
+/*
+ * Return column-major index from 2d array index, A[row,col].
+ */
+template<class I>
+inline I col_major(const I row, const I col, const I num_rows)
+{
+    return col*num_rows + row;
+}
+
+
 /* dot(x, y, n)
  *
  * Parameters
@@ -176,6 +196,11 @@ template<class I, class T, class F>
 inline void norm(const T x[], const I n, F &normx)
 {
     normx = sqrt(real(dot_prod(x,x,n)));
+}
+template<class I, class T>
+inline T norm(const T x[], const I n)
+{
+    return std::sqrt(dot_prod(x,x,n));
 }
 
 
@@ -447,7 +472,9 @@ inline void gemm(const T Ax[], const I Arows, const I Acols, const char Atrans,
             }
         }
     }
-
+    else {
+        std::cout << "Unsupported combination of row/column major for dense multiplication.\n";
+    }
 }
 
 
@@ -507,7 +534,6 @@ inline void gemm(const T Ax[], const I Arows, const I Acols, const char Atrans,
  * Vol 10, No 2, p 359-371, March 1989.
  *
  */
-
 template<class I, class T, class F>
 I svd_jacobi(const T Ax[], T Tx[], T Bx[], F Sx[], const I m, const I n)
 {
@@ -845,6 +871,7 @@ void svd_solve( T Ax[], I m, I n, T b[], F sing_vals[], T work[], I work_size)
     return;
 }
 
+
 /* Replace each block of A with a Moore-Penrose pseudoinverse of that block.
  * Routine is designed to invert many small matrices at once.
  *
@@ -958,6 +985,7 @@ void pinv_array(T AA[], const int AA_size,
     return;
 }
 
+
 /*
  * Scale the columns of a CSC matrix *in place*
  *
@@ -984,6 +1012,7 @@ void csc_scale_columns(const I n_row,
     }
 }
 
+
 /*
  * Scale the rows of a CSC matrix *in place*
  *
@@ -1006,6 +1035,424 @@ void csc_scale_rows(const I n_row,
     const I nnz = Ap[n_col];
     for(I i = 0; i < nnz; i++){
         Ax[i] *= Xx[Aj[i]];
+    }
+}
+
+
+/* QR-decomposition using Householer transformations on dense
+ * 2d array stored in either column- or row-major form.
+ *
+ * Parameters
+ * ----------
+ *        A : double array
+ *            2d matrix A stored in 1d column- or row-major.
+ *        m : &int
+ *            Number of rows in A
+ *        n : &int
+ *            Number of columns in A
+ *        is_col_major : bool
+ *            True if A is stored in column-major, false
+ *            if A is stored in row-major.
+ *
+ * Returns
+ * -------
+ *        Q : vector<double>
+ *            Matrix Q stored in same format as A.
+ *        R : in-place
+ *            R is stored over A in place, in same format.
+ *
+ * Notes
+ * ------
+ * Currently only set up for real-valued matrices. May easily
+ * generalize to complex, but haven't checked.
+ *
+ */
+template<class I, class T>
+std::vector<T> QR(T A[],
+                  const I &m,
+                  const I &n,
+                  const I is_col_major)
+{
+    // Funciton pointer for row or column major matrices
+    I (*get_ind)(const I, const I, const I);
[…]
