+++ pyamg/amg_core/linalg.h
@@ -1044,34 +1044,33 @@ void csc_scale_rows(const I n_row,
  *
  * Parameters
  * ----------
- *        A : double array
- *            2d matrix A stored in 1d column- or row-major.
- *        m : &int
- *            Number of rows in A
- *        n : &int
- *            Number of columns in A
- *        is_col_major : bool
- *            True if A is stored in column-major, false
- *            if A is stored in row-major.
+ * A : double array
+ *     2d matrix A stored in 1d column- or row-major.
+ * m : int
+ *     Number of rows in A
+ * n : int
+ *     Number of columns in A
+ * is_col_major : bool
+ *     True if A is stored in column-major, false
+ *     if A is stored in row-major.
  *
  * Returns
  * -------
- *        Q : vector<double>
- *            Matrix Q stored in same format as A.
- *        R : in-place
- *            R is stored over A in place, in same format.
+ * Q : vector<double>
+ *     Matrix Q stored in same format as A.
+ * R : in-place
+ *     R is stored over A in place, in same format.
  *
  * Notes
- * ------
- * Currently only set up for real-valued matrices. May easily
- * generalize to complex, but haven't checked.
+ * -----
+ * Currently only set up for real-valued matrices.
  *
  */
 template<class I, class T>
-std::vector<T> QR(T A[],
+std::vector<T> qr(T A[],
                   const I &m,
                   const I &n,
-                  const I is_col_major)
+                  const bool is_col_major)
 {
     // Funciton pointer for row or column major matrices
     I (*get_ind)(const I, const I, const I);
@@ -1166,23 +1165,23 @@ std::vector<T> QR(T A[],
  *
  * Parameters
  * ----------
- *        R : double array, length m*n
- *            Upper-triangular array stored in column- or row-major.
- *        rhs : double array, length m
- *            Right hand side of linear system
- *        x : double array, length n
- *            Preallocated array for solution
- *        m : &int
- *            Number of rows in R
- *        n : &int
- *            Number of columns in R
- *        is_col_major : bool
- *            True if R is stored in column-major, false
- *            if R is stored in row-major.
+ * R : double array, length m*n
+ *     Upper-triangular array stored in column- or row-major.
+ * rhs : double array, length m
+ *     Right hand side of linear system
+ * x : double array, length n
+ *     Preallocated array for solution
+ * m : int
+ *     Number of rows in R
+ * n : int
+ *     Number of columns in R
+ * is_col_major : bool
+ *     True if R is stored in column-major, false
+ *     if R is stored in row-major.
  *
  * Returns
  * -------
- *        Nothing, solution is stored in x[].
+ * Nothing, solution is stored in x[].
  *
  * Notes
  * -----
@@ -1198,7 +1197,7 @@ void upper_tri_solve(const T R[],
                      T x[],
                      const I m,
                      const I n,
-                     const I is_col_major)
+                     const bool is_col_major)
 {
     // Function pointer for row or column major matrices
     I (*get_ind)(const I, const I, const I);
@@ -1239,23 +1238,23 @@ void upper_tri_solve(const T R[],
  *
  * Parameters
  * ----------
- *        L : double array, length m*n
- *            Lower-triangular array stored in column- or row-major.
- *        rhs : double array, length m
- *            Right hand side of linear system
- *        x : double array, length n
- *            Preallocated array for solution
- *        m : &int
- *            Number of rows in L
- *        n : &int
- *            Number of columns in L
- *        is_col_major : bool
- *            True if L is stored in column-major, false
- *            if L is stored in row-major.
[…]
