### 2022-03-09 1e96db3
docs: fix docstring
