### 2022-03-09 2edbce2
refactor: limit to only filter rows
