+++ pyamg/amg_core/linalg.h
@@ -128,26 +128,6 @@ inline std::complex<double> zero_imag(std::complex<double>& x)
  *      templated for pyamg's complex class, float and double
  *******************************************************************/
 
-/*
- * Return row-major index from 2d array index, A[row,col].
- */
-template<class I>
-inline I row_major(const I row, const I col, const I num_cols)
-{
-    return row*num_cols + col;
-}
-
-
-/*
- * Return column-major index from 2d array index, A[row,col].
- */
-template<class I>
-inline I col_major(const I row, const I col, const I num_rows)
-{
-    return col*num_rows + row;
-}
-
-
 /* dot(x, y, n)
  *
  * Parameters
@@ -183,6 +163,8 @@ inline T dot_prod(const T x[], const T y[], const I n)
  *     n-vector
  * n : int
  *     size of x and y
+ * normx : float
+ *     output value
  *
  * Returns
  * -------
@@ -194,11 +176,6 @@ template<class I, class T, class F>
 inline void norm(const T x[], const I n, F &normx)
 {
     normx = sqrt(real(dot_prod(x,x,n)));
-}
-template<class I, class T>
-inline T norm(const T x[], const I n)
-{
-    return std::sqrt(dot_prod(x,x,n));
 }
 
 
@@ -869,7 +846,6 @@ void svd_solve( T Ax[], I m, I n, T b[], F sing_vals[], T work[], I work_size)
     return;
 }
 
-
 /* Replace each block of A with a Moore-Penrose pseudoinverse of that block.
  * Routine is designed to invert many small matrices at once.
  *
@@ -983,7 +959,6 @@ void pinv_array(T AA[], const int AA_size,
     return;
 }
 
-
 /*
  * Scale the columns of a CSC matrix *in place*
  *
@@ -1010,7 +985,6 @@ void csc_scale_columns(const I n_row,
     }
 }
 
-
 /*
  * Scale the rows of a CSC matrix *in place*
  *
@@ -1034,337 +1008,6 @@ void csc_scale_rows(const I n_row,
     for(I i = 0; i < nnz; i++){
         Ax[i] *= Xx[Aj[i]];
     }
-}
-
-
-/* QR-decomposition using Householer transformations on dense
- * 2d array stored in either column- or row-major form.
- *
- * Parameters
- * ----------
- * A : double array
- *     2d matrix A stored in 1d column- or row-major.
- * m : int
- *     Number of rows in A
- * n : int
- *     Number of columns in A
- * is_col_major : bool
- *     True if A is stored in column-major, false
- *     if A is stored in row-major.
- *
- * Returns
- * -------
- * Q : vector<double>
- *     Matrix Q stored in same format as A.
- * R : in-place
- *     R is stored over A in place, in same format.
- *
- * Notes
- * -----
- * Currently only set up for real-valued matrices.
- *
- */
-template<class I, class T>
-std::vector<T> qr(T A[],
-                  const I &m,
-                  const I &n,
-                  const bool is_col_major)
-{
-    // Funciton pointer for row or column major matrices
-    I (*get_ind)(const I, const I, const I);
-    const I *C;
-    if (is_col_major) {
-        get_ind = &col_major;
-        C = &m;
-    }
-    else {
-        get_ind = &row_major;
-        C = &n;
-    }
-
-    // Initialize Q to identity
-    std::vector<T> Q(m*m,0);
-    for (I i=0; i<m; i++) {
-        Q[get_ind(i,i,m)] = 1;
-    }
-
-    // Loop over columns of A using Householder reflections
-    for (I j=0; j<n; j++) {
-
-        // Break loop for short fat matrices
-        if (m <= j) {
-            break;
-        }
-
-        // Get norm of next column of A to be reflected. Choose sign
[…]
