### 2022-03-09 18db6e3
fix: adjust lumping in complex case to ignore diagonal
