### 2022-03-09 46bfdab
fix: adjust lumping in complex case to ignore diagonal
