+++ pyamg/tests/test_amg_core_linalg.py
@@ -1,0 +1,93 @@
+import numpy as np
+from scipy import sparse
+from pyamg import amg_core
+
+def test_real():
+    A = np.array([[ 1. ,  2. ,  0.5,  0. ],
+                  [ 1. ,  2. ,  1.5,  3. ],
+                  [ 0. ,  4. ,  4. ,  2.1],
+                  [-5. ,  2. ,  4. , 10. ]])
+    A0 = A.copy()
+
+    # no lumping, threhold 0.0
+    A = sparse.csr_matrix(A0.copy())
+    amg_core.linalg.filter_matrix_rows(4, 0.0, A.indptr, A.indices, A.data, 0)
+    np.testing.assert_array_equal(A.toarray(), A0)
+
+    # no lumping, threhold 1.0
+    A = sparse.csr_matrix(A0.copy())
+    amg_core.linalg.filter_matrix_rows(4, 1.0, A.indptr, A.indices, A.data, 0)
+    B = np.array([[ 1. ,  2. ,  0.0,  0. ],
+                  [ 0. ,  2. ,  0.0,  3. ],
+                  [ 0. ,  4. ,  4. ,  0.0],
+                  [ 0. ,  0. ,  0. , 10. ]])
+    np.testing.assert_array_equal(A.toarray(), B)
+
+    # lumping, threhold 1.0
+    A = sparse.csr_matrix(A0.copy())
+    amg_core.linalg.filter_matrix_rows(4, 1.0, A.indptr, A.indices, A.data, True)
+    B = np.array([[ 1.5,  2. ,  0.0,  0. ],
+                  [ 0. ,  4.5,  0.0,  3. ],
+                  [ 0. ,  4. ,  6.1,  0.0],
+                  [ 0. ,  0. ,  0. , 11. ]])
+    np.testing.assert_array_equal(A.toarray(), B)
+
+def test_imag():
+    A = np.array([[ 1. ,  2. ,  0.5,  0. ],
+                  [ 1. ,  2. ,  1.5,  3. ],
+                  [ 0. ,  4. ,  4. ,  2.1],
+                  [-5. ,  2. ,  4. , 10. ]]) * 1j
+    A0 = A.copy()
+
+    # no lumping, threhold 0.0
+    A = sparse.csr_matrix(A0.copy())
+    amg_core.linalg.filter_matrix_rows(4, 0.0, A.indptr, A.indices, A.data, 0)
+    np.testing.assert_array_equal(A.toarray(), A0)
+
+    # no lumping, threhold 1.0
+    A = sparse.csr_matrix(A0.copy())
+    amg_core.linalg.filter_matrix_rows(4, 1.0, A.indptr, A.indices, A.data, 0)
+    B = np.array([[ 1. ,  2. ,  0.0,  0. ],
+                  [ 0. ,  2. ,  0.0,  3. ],
+                  [ 0. ,  4. ,  4. ,  0.0],
+                  [ 0. ,  0. ,  0. , 10. ]]) * 1j
+    np.testing.assert_array_equal(A.toarray(), B)
+
+    # lumping, threhold 1.0
+    A = sparse.csr_matrix(A0.copy())
+    amg_core.linalg.filter_matrix_rows(4, 1.0, A.indptr, A.indices, A.data, True)
+    B = np.array([[ 1.5,  2. ,  0.0,  0. ],
+                  [ 0. ,  4.5,  0.0,  3. ],
+                  [ 0. ,  4. ,  6.1,  0.0],
[…]
