+++ pyamg/util/utils.py
@@ -2056,37 +2056,38 @@ def filter_matrix_rows(A, theta, diagonal=False, lump=False):
         A.eliminate_zeros()
         if Aformat == 'bsr':
             A = A.tobsr(blocksize=blocksize)
-        return
+        # no return, inplace
 
-    # Apply drop-tolerance to each row of A.  We apply the drop-tolerance with
-    # amg_core.classical_strength_of_connection_abs(), which ignores diagonal
-    # entries, thus necessitating the trick where we add A.shape[0] to each of
-    # the row indices
-    A_filter = A.copy()
-    A.indices += A.shape[0]
-    A_filter.indices += A.shape[0]
-
-    # classical_strength_of_connection takes an absolute value internally
-    amg_core.classical_strength_of_connection_abs(A.shape[0],
-                                                  theta,
-                                                  A.indptr,
-                                                  A.indices,
-                                                  A.data,
-                                                  A_filter.indptr,
-                                                  A_filter.indices,
-                                                  A_filter.data)
-    A_filter.indices[:A_filter.indptr[-1]] -= A_filter.shape[0]
-    A_filter = csr_matrix((A_filter.data[:A_filter.indptr[-1]],
-                           A_filter.indices[:A_filter.indptr[-1]],
-                           A_filter.indptr), shape=A_filter.shape)
-
-    if Aformat == 'bsr':
-        A_filter = A_filter.tobsr(blocksize)
     else:
-        A_filter = A_filter.asformat(Aformat)
+        # Apply drop-tolerance to each row of A.  We apply the drop-tolerance with
+        # amg_core.classical_strength_of_connection_abs(), which ignores diagonal
+        # entries, thus necessitating the trick where we add A.shape[0] to each of
+        # the row indices
+        A_filter = A.copy()
+        A.indices += A.shape[0]
+        A_filter.indices += A.shape[0]
 
-    A.indices -= A.shape[0]
-    return A_filter
+        # classical_strength_of_connection takes an absolute value internally
+        amg_core.classical_strength_of_connection_abs(A.shape[0],
+                                                      theta,
+                                                      A.indptr,
+                                                      A.indices,
+                                                      A.data,
+                                                      A_filter.indptr,
+                                                      A_filter.indices,
+                                                      A_filter.data)
+        A_filter.indices[:A_filter.indptr[-1]] -= A_filter.shape[0]
+        A_filter = csr_matrix((A_filter.data[:A_filter.indptr[-1]],
+                               A_filter.indices[:A_filter.indptr[-1]],
+                               A_filter.indptr), shape=A_filter.shape)
+
+        if Aformat == 'bsr':
+            A_filter = A_filter.tobsr(blocksize)
+        else:
+            A_filter = A_filter.asformat(Aformat)
+
+        A.indices -= A.shape[0]
+        return A_filter
 
 
 def truncate_rows(A, nz_per_row):
