+++ pyamg/amg_core/__init__.py
@@ -12,7 +12,7 @@ from .graph import (maximal_independent_set_serial, maximal_independent_set_para
                     breadth_first_search, connected_components)
 
 from .krylov import (apply_householders, householder_hornerscheme, apply_givens)
-from .linalg import (pinv_array, csc_scale_columns, csc_scale_rows)
+from .linalg import (pinv_array, csc_scale_columns, csc_scale_rows, filter_matrix_rows)
 from .relaxation import (gauss_seidel, bsr_gauss_seidel, gauss_seidel_indexed,
                          jacobi, bsr_jacobi,
                          jacobi_ne, gauss_seidel_ne, gauss_seidel_nr,
@@ -59,6 +59,7 @@ __all__ = [
     'pinv_array',
     'csc_scale_columns',
     'csc_scale_rows',
+    'filter_matrix_rows',
     #
     'gauss_seidel',
     'bsr_gauss_seidel',
