+++ README.md
@@ -50,7 +50,7 @@ For documentation see [http://pyamg.readthedocs.io/en/latest/](http://pyamg.read
 
 Create an [issue](https://github.com/pyamg/pyamg/issues).
 
-Look at the [Tutorial](https://github.com/pyamg/pyamg/wiki/Tutorial) or the [Examples](https://github.com/pyamg/pyamg/wiki/Examples) (for instance  the [0STARTHERE](https://github.com/pyamg/pyamg-examples/blob/master/0STARTHERE/demo.py) example).
+Look at the [Tutorial](https://github.com/pyamg/pyamg/wiki/Tutorial) or the [examples](https://github.com/pyamg/pyamg-examples) (for instance  the [0_start_here](https://github.com/pyamg/pyamg-examples/blob/main/0_start_here/demo.py) example).
 
 # What is AMG?
 
