### 2022-03-17 pyamg/pyamg#321 @mayrmt
Review for JOSS publication: contribution and authorship
While reviewing the manuscript about PyAMG submitted to JOSS (https://github.com/openjournals/joss-reviews/issues/4142), I stumbled across the following items:

**Authorship:** The list of authors of the accompanying PDF submitted to JOSS lists Luke Olsen and Jacob Schroder. The file [`DEVELOPERS.txt`](https://github.com/pyamg/pyamg/blob/main/DEVELOPERS.txt) however lists also Nathan Bell.

@lukeolson Please elaborate on this differences in the lists of authors/developers.
