### 2022-03-17 #4142 @mayrmt
Some remarks on my assessment:

## General checks

- Contribution and authorship: see
   - Contribution: The submitting author @lukeolsen has made major contributions to the software.
   - Authorship: see https://github.com/pyamg/pyamg/issues/321

## Functionality

I have verified all listed items on my machine. ✅

## Documentation

- Options to look for documentation and support are outlined in `README.md`
- Basic information on contributing to the source code is given in `CONTRIBUTING.md`
- Automated tests are part of pull requests to the source code repository.

## Software Paper

- A statement of need: The paper does not contain a section entitled **"Statement of Need"**. Yet, it becomes clear from the **"Summary"** section, what problems the software is designed to solve and who the target audience is. For me, this is sufficient, though if the journal guidelines require such a section, then of course it needs to be added. @jedbrown Can you give advice on the journal guidelines here? Thanks in advance?
- State of the field:
   - The **"Summary"** section sheds some light on the relation to established Python implementations for solving linear systems, in particular with direct or iterative methods, and it also explains, how the contribution of PyAMG fits in to the context.
   - A statement on the relation to other AMG software packages (such as Hypre, Trilinos/MueLu, PETSc, ...) is missing.
