### 2022-03-21 d6c7618
docs: fix spacing in docstrings
