+++ pyamg/amg_core/sparse.h
@@ -1,3 +1,4 @@
+#include <complex>
 //
 // Threaded SpMV
 //
