### 2022-03-22 6ea1fb1
add potential mac omp flag test
