### 2022-03-22 9fec26a
update mac flags
