### 2022-03-22 e289f23
fix omp_str
