### 2022-03-22 18457c0
trying automatic paths
