+++ setup.py
@@ -26,7 +26,8 @@ amg_core_headers = ['air',
                     'linalg',
                     'relaxation',
                     'ruge_stuben',
-                    'smoothed_aggregation']
+                    'smoothed_aggregation',
+                    'sparse']
 
 ext_modules = [
     Pybind11Extension(f'pyamg.amg_core.{f}',
