+++ setup.py
@@ -17,7 +17,22 @@ supporting C++ code for performance critical operations.
 """
 
 from setuptools import setup
-from pybind11.setup_helpers import Pybind11Extension, build_ext
+from pybind11.setup_helpers import Pybind11Extension, build_ext, has_flag
+
+extra_compile_args = []
+extra_link_args = []
+
+for cflag, lflag in [('-fopenmp', '-fopenmp'),
+                     ('-Xclang -fopenmp', '-lomp'),
+                     ('-Xpreprocessor -fopenmp -lomp', '-lomp')] :
+    try:
+        openmp = has_flag(build_ext.compiler, cflag)
+        extra_compile_args.append(cflag)
+        extra_link_args.append(lflag)
+    except:
+        openmp = False
+
+print(f'\n\nUsing openmp flag {extra_compile_args}\n\n')
 
 amg_core_headers = ['air',
                     'evolution_strength',
@@ -31,6 +46,8 @@ amg_core_headers = ['air',
 
 ext_modules = [
     Pybind11Extension(f'pyamg.amg_core.{f}',
+                      extra_compile_args=extra_compile_args,
+                      extra_link_args=extra_link_args,
                       sources=[f'pyamg/amg_core/{f}_bind.cpp'],
                      )
     for f in amg_core_headers]
