### 2022-03-22 pyamg/pyamg#321 @mayrmt
Has been addressed in the manuscript draft. Closing.
