### 2022-03-22 #4142 @lukeolson
@mayrmt many thanks for your time and for your review.

> * A statement of need: The paper does not contain a section entitled **"Statement of Need"**. Yet, it becomes clear from the **"Summary"** section, what problems the software is designed to solve and who the target audience is. For me, this is sufficient, though if the journal guidelines require such a section, then of course it needs to be added. @jedbrown Can you give advice on the journal guidelines here? Thanks in advance?

Let's let @jedbrown weigh in on a separate section for a Statement of Need -- happy to add one if it's needed.

> * State of the field:
>
>   * A statement on the relation to other AMG software packages (such as Hypre, Trilinos/MueLu, PETSc, ...) is missing.

Yep, fair point.  We can add a blurb on this -- good idea.
