### 2022-03-22 275e47e
add blurb on other AMG packages
