+++ docs/paper/paper.bib
@@ -108,3 +108,103 @@
   year    = 2006,
   month   = Oct,
 }
+
+@article{henson2002155,
+  title = {{B}oomer{AMG}: A parallel algebraic multigrid solver and preconditioner},
+  journal = {Applied Numerical Mathematics},
+  volume = {41},
+  number = {1},
+  pages = {155-177},
+  year = {2002},
+  note = {Developments and Trends in Iterative Methods for Large Systems of Equations - in memorium Rudiger Weiss},
+  issn = {0168-9274},
+  doi = {https://doi.org/10.1016/S0168-9274(01)00115-5},
+  url = {https://www.sciencedirect.com/science/article/pii/S0168927401001155},
+  author = {Van Emden Henson and Ulrike Meier Yang},
+  keywords = {Algebraic multigrid, Parallel computing},
+}
+
+@misc{hypre,
+  title = {hypre: high performance preconditioners},
+  url = {www.llnl.gov/casc/hypre/},
+  year = {2022},
+}
+
+@manual{trilinos-website,
+  title = {The {T}rilinos {P}roject {W}ebsite},
+  author = {The {T}rilinos {P}roject {T}eam}},
+  year = {2020 (acccessed May 22, 2020)},
+  url = {https://trilinos.github.io}
+}
+
+@manual{muelu-website,
+  title = {The {M}uelu {P}roject {W}ebsite},
+  author = {The {M}uelu {P}roject {T}eam}},
+  year = {2020 (acccessed May 22, 2020)},
+  url = {https://trilinos.github.io/muelu.html}
+}
+
+@misc{petsc-web-page,
+  author = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed Brown
+    and Peter Brune and Kris Buschelman and Emil~M. Constantinescu and Lisandro Dalcin and Alp Dener
+    and Victor Eijkhout and William~D. Gropp and V\'{a}clav Hapla and Tobin Isaac and Pierre Jolivet
+    and Dmitry Karpeev and Dinesh Kaushik and Matthew~G. Knepley and Fande Kong and Scott Kruger
+    and Dave~A. May and Lois Curfman McInnes and Richard Tran Mills and Lawrence Mitchell and Todd Munson
+    and Jose~E. Roman and Karl Rupp and Patrick Sanan and Jason Sarich and Barry~F. Smith
+    and Stefano Zampini and Hong Zhang and Hong Zhang and Junchao Zhang},
+  title        = {{PETS}c {W}eb page},
+  url          = {https://petsc.org/},
+  howpublished = {\url{https://petsc.org/}},
+  year         = {2021},
+}
+
+@MISC{cusp,
+  author = "Steven Dalton and Nathan Bell and Luke Olson and Michael Garland",
+  title = "Cusp: Generic Parallel Algorithms for Sparse Matrix and Graph Computations",
+  year = "2014",
+  url = "http://cusplibrary.github.io/",
+  note = "Version 0.5.0"
+}
+
+@article{amgx,
+  author = {Naumov, M. and Arsaev, M. and Castonguay, P. and Cohen, J. and Demouth, J. and Eaton, J. and Layton, S. and Markovskiy, N. and Reguly, I. and Sakharnykh, N. and Sellappan, V. and Strzodka, R.},
+  title = {{AmgX}: A Library for {GPU} Accelerated Algebraic Multigrid and Preconditioned Iterative Methods},
+  journal = {SIAM Journal on Scientific Computing},
+  volume = {37},
+  number = {5},
+  pages = {S602-S626},
+  year = {2015},
+  doi = {10.1137/140980260},
[…]
