### 2022-03-22 7e42f15
remove errant brace in bib file
