+++ docs/paper/paper.bib
@@ -61,7 +61,6 @@
   url        = {https://lukeo.cs.illinois.edu/files/2015_Ol_encamg.pdf},
 }
 
-
 @incollection{encmg,
   author    = {Olson, Luke},
   editor    = {Engquist, Bj{\"o}rn},
@@ -125,9 +124,10 @@
 }
 
 @misc{hypre,
-  title = {hypre: high performance preconditioners},
-  url   = {www.llnl.gov/casc/hypre/},
-  year  = {2022},
+  author = {},
+  title  = {hypre: high performance preconditioners},
+  url    = {https://github.com/hypre-space/hypre},
+  year   = {2022},
 }
 
 @manual{trilinos-website,
@@ -158,7 +158,7 @@
   year         = {2021},
 }
 
-@MISC{cusp,
+@misc{cusp,
   author = {Steven Dalton and Nathan Bell and Luke Olson and Michael Garland},
   title  = {Cusp: Generic Parallel Algorithms for Sparse Matrix and Graph Computations},
   year   = {2014},
