+++ docs/paper/paper.bib
@@ -124,8 +124,8 @@
 }
 
 @misc{hypre,
-  author = {},
-  title  = {hypre: high performance preconditioners},
+  author = {hypre},
+  title  = {high performance preconditioners},
   url    = {https://github.com/hypre-space/hypre},
   year   = {2022},
 }
