+++ docs/paper/paper.bib
@@ -133,14 +133,14 @@
 @manual{trilinos-website,
   title  = {The {T}rilinos {P}roject {W}ebsite},
   author = {The {T}rilinos {P}roject {T}eam},
-  year   = {2020 (acccessed May 22, 2020)},
+  year   = {2020},
   url    = {https://trilinos.github.io}
 }
 
 @manual{muelu-website,
   title  = {The {M}uelu {P}roject {W}ebsite},
   author = {The {M}uelu {P}roject {T}eam},
-  year   = {2020 (acccessed May 22, 2020)},
+  year   = {2020},
   url    = {https://trilinos.github.io/muelu.html}
 }
 
