### 2022-03-23 #4142 @lukeolson
> >   * A statement on the relation to other AMG software packages (such as Hypre, Trilinos/MueLu, PETSc, ...) is missing.
>
> Yep, fair point. We can add a blurb on this -- good idea.

@mayrmt Added a couple sentences at the end of the summary section.
