### 2022-03-23 cc12a5e
update setup; fix test/trys in scripts
