+++ scripts/try_omp.py
@@ -2,7 +2,7 @@ import pyamg
 import numpy as np
 from timeit import default_timer as timer
 
-n = 300
+n = 200
 
 t0 = timer()
 A = pyamg.gallery.poisson((n,n,n), format='csr')
@@ -10,14 +10,19 @@ ml = pyamg.smoothed_aggregation_solver(A, max_coarse=10)
 b = np.random.rand(A.shape[0])
 u = np.random.rand(A.shape[0])
 t1 = timer()
-print('setup time {}'.format(t1-t0))
+print(f'problem size={A.shape}')
+print(f'setup time {t1-t0}')
 
 t0 = timer()
-x = ml.solve(b, x0=u)
+res = []
+x = ml.solve(b, x0=u, residuals=res)
 t1 = timer()
-print('reg time {}'.format(t1-t0))
+print(f'reg time {t1-t0}')
+print(res, '\n')
 
 t0 = timer()
-x = ml.solve(b, x0=u, openmp=True)
+res = []
+x = ml.solve(b, x0=u, openmp=True, residuals=res)
 t1 = timer()
-print('omp time {}'.format(t1-t0))
+print(f'omp time {t1-t0}')
+print(res, '\n')
