### 2022-03-23 c46d26e
lint: remove unneeded imports.  fix docstrings
