### 2022-03-23 2435c06
updating OMP discussion in two files
