### 2022-03-25 #4142 @mattmartineau
### General checks

**Substantial scholarly effort:** It meets the criteria laid out in the guidelines and I think it is an novel project. From my perspective, it serves as a very interesting platform for demonstrating the range of AMG methods and may even be useful as a starting point for my own work (we want to implement smoothed aggregation in AmgX in the near future).

### Functionality

**Functionality:** The instructions were easy to follow and I was able to match the output of several examples immediately.

**Performance:** There isn't a direct claim regarding performance that I came across, except that performance is a key goal of the project, and I feel that it passes this point of the review. I'll note that I found the recent work on OpenMP parallelisation (in the sparse-ops project), and that seems like a great step for runtime performance. I think it is important, in general, that parallelism is being addressed long term due to the impact on implementation and optimal configuration when applied to real-world academic/industrial problems.

### Documentation

**Functionality documentation:** I think the documentation is at a good level of granularity for the likely intended audience. Interfaces are well documented and referenced, often offering example code with example outputs.

### Software paper

**A statement of need:** I personally think that the "Summary" section very adequately explains why this project is distinct from other OSS projects, and motivates it's existence. As previously mentioned, it strictly doesn't have that separate section, however the current structure works well as the points are well contextualised.
