### 2022-03-25 #4142 @lukeolson
Thanks @mattmartineau for your time on this and for going over it in detail (indeed [omp](https://github.com/pyamg/pyamg/pull/324) is on the way... at least for the solve phase).

We'll wait to hear from @jedbrown to see if we can leave the Summary as-is.  Any other changes, just let us know.
