### 2022-03-27 #4142 @jedbrown
Thanks (just getting back from a few days of desert climbing with the kids). I agree that a summary articulating this should be sufficient, but there is an automated check in the acceptance process looking for a section heading named "Statement of need". (This check got a little stricter last month and I'm checking on whether that was deliberate.)
