### 2022-04-02 #4142 @jedbrown
Sorry about my slow reply. If it's no big hassle for you, naming a section `Statement of need` is definitely the easiest. If you don't like that style, I think we can make an exception (and perhaps a slight adjustment of policy, but that may take longer).
