+++ pyamg/krylov/_gmres_mgs.py
@@ -35,12 +35,22 @@ def apply_givens(Q, v, k):
     """
     for j in range(k):
         Qloc = Q[j]
-        v[j:j+2] = np.dot(Qloc, v[j:j+2])
+        v[j : j + 2] = np.dot(Qloc, v[j : j + 2])
 
 
-def gmres_mgs(A, b, x0=None, tol=1e-5,
-              restrt=None, maxiter=None,
-              M=None, callback=None, residuals=None, reorth=False):
+def gmres_mgs(
+    A,
+    b,
+    x0=None,
+    tol=1e-5,
+    restrt=None,
+    maxiter=None,
+    M=None,
+    H=None,
+    callback=None,
+    residuals=None,
+    reorth=False,
+):
     """Generalized Minimum Residual Method (GMRES) based on MGS.
 
     GMRES iteratively refines the initial solution guess to the system
@@ -72,6 +82,8 @@ def gmres_mgs(A, b, x0=None, tol=1e-5,
         - defaults to min(n,40) if restart=None
     M : array, matrix, sparse matrix, LinearOperator
         n x n, inverted preconditioner, i.e. solve M A x = M b.
+    H : array, matrix, sparse matrix, LinearOperator
+        n x n, Hermitian positive definite matrix induced by the H-norm GMRES.
     callback : function
         User-supplied function is called after each iteration as
         callback(xk), where xk is the current solution vector
@@ -129,22 +141,35 @@ def gmres_mgs(A, b, x0=None, tol=1e-5,
 
     """
     # Convert inputs to linear system, with error checking
+    if H is not None:
+        H = make_system(A, H, x0, b)[1]
     A, M, x, b, postprocess = make_system(A, M, x0, b)
     n = A.shape[0]
 
     # Ensure that warnings are always reissued from this function
-    warnings.filterwarnings('always', module='pyamg.krylov._gmres_mgs')
+    warnings.filterwarnings("always", module="pyamg.krylov._gmres_mgs")
 
     # Get fast access to underlying BLAS routines
     # dotc is the conjugate dot, dotu does no conjugation
-    [lartg] = get_lapack_funcs(['lartg'], [x])
+    [lartg] = get_lapack_funcs(["lartg"], [x])
     if np.iscomplexobj(np.zeros((1,), dtype=x.dtype)):
-        [axpy, dotu, dotc, scal] =\
-            get_blas_funcs(['axpy', 'dotu', 'dotc', 'scal'], [x])
+        [axpy, dotu, dotc, scal] = get_blas_funcs(["axpy", "dotu", "dotc", "scal"], [x])
     else:
         # real type
-        [axpy, dotu, dotc, scal] =\
-            get_blas_funcs(['axpy', 'dot', 'dot', 'scal'], [x])
+        [axpy, dotu, dotc, scal] = get_blas_funcs(["axpy", "dot", "dot", "scal"], [x])
+
+    if H is not None:
+        # Define the H-norm and H-inner product
+        def dotcH(x, y):
+            temp = H @ y
+            return dotc(x, temp)
+
+        def normH(x):
+            return np.sqrt(dotcH(x, x).real)
+
+    else:
+        dotcH = dotc
+        normH = norm
 
     # Set number of outer and inner iterations
     # If no restarts,
@@ -157,7 +182,7 @@ def gmres_mgs(A, b, x0=None, tol=1e-5,
         else:
             max_outer = 1
         if restrt > n:
-            warn('Setting restrt to maximum allowed, n.')
+            warn("Setting restrt to maximum allowed, n.")
             restrt = n
         max_inner = restrt
     else:
@@ -165,14 +190,14 @@ def gmres_mgs(A, b, x0=None, tol=1e-5,
[…]
