+++ pyamg/krylov/_gmres_householder.py
@@ -18,9 +18,17 @@ def _mysign(x):
     return x / np.abs(x)
 
 
-def gmres_householder(A, b, x0=None, tol=1e-5,
-                      restrt=None, maxiter=None,
-                      M=None, callback=None, residuals=None):
+def gmres_householder(
+    A,
+    b,
+    x0=None,
+    tol=1e-5,
+    restrt=None,
+    maxiter=None,
+    M=None,
+    callback=None,
+    residuals=None,
+):
     """Generalized Minimum Residual Method (GMRES) based on Housholder.
 
     GMRES iteratively refines the initial solution guess to the
@@ -111,10 +119,10 @@ def gmres_householder(A, b, x0=None, tol=1e-5,
     n = A.shape[0]
 
     # Ensure that warnings are always reissued from this function
-    warnings.filterwarnings('always', module='pyamg.krylov._gmres_householder')
+    warnings.filterwarnings("always", module="pyamg.krylov._gmres_householder")
 
     # Get fast access to underlying LAPACK routine
-    [lartg] = get_lapack_funcs(['lartg'], [x])
+    [lartg] = get_lapack_funcs(["lartg"], [x])
 
     # Set number of outer and inner iterations
     # If no restarts,
@@ -127,7 +135,7 @@ def gmres_householder(A, b, x0=None, tol=1e-5,
         else:
             max_outer = 1
         if restrt > n:
-            warn('Setting restrt to maximum allowed, n.')
+            warn("Setting restrt to maximum allowed, n.")
             restrt = n
         max_inner = restrt
     else:
@@ -135,14 +143,14 @@ def gmres_householder(A, b, x0=None, tol=1e-5,
         if maxiter is None:
             maxiter = min(n, 40)
         elif maxiter > n:
-            warn('Setting maxiter to maximum allowed, n.')
+            warn("Setting maxiter to maximum allowed, n.")
             maxiter = n
         max_inner = maxiter
 
     # Is this a one dimensional matrix?
     if n == 1:
         entry = np.ravel(A @ np.array([1.0], dtype=x.dtype))
-        return (postprocess(b/entry), 0)
+        return (postprocess(b / entry), 0)
 
     # Prep for method
     r = b - A @ x
@@ -189,7 +197,7 @@ def gmres_householder(A, b, x0=None, tol=1e-5,
         # upper Hessenberg matrix (made upper tri with Givens Rotations)
         H = np.zeros((max_inner, max_inner), dtype=x.dtype)
         # Householder reflectors
-        W = np.zeros((max_inner+1, n), dtype=x.dtype)
+        W = np.zeros((max_inner + 1, n), dtype=x.dtype)
         W[0, :] = w
 
         # Multiply r with (I - 2*w*w.T), i.e. apply the Householder reflector
@@ -205,7 +213,7 @@ def gmres_householder(A, b, x0=None, tol=1e-5,
             # (2) Calculate the rest, v = P_1*P_2*P_3...P_{j-1}*ej.
             # for j in range(inner-1,-1,-1):
             #    v -= 2.0*dot(conjugate(W[j,:]), v)*W[j,:]
-            amg_core.apply_householders(v, np.ravel(W), n, inner-1, -1, -1)
+            amg_core.apply_householders(v, np.ravel(W), n, inner - 1, -1, -1)
 
             # Calculate new search direction
             v = A @ v
@@ -221,7 +229,7 @@ def gmres_householder(A, b, x0=None, tol=1e-5,
             # direction
             # for j in range(inner+1):
             #    v -= 2.0*dot(conjugate(W[j,:]), v)*W[j,:]
[…]
