### 2022-04-06 bbed48a
Added gen AIR code w/ auxiliary coarse solve
