+++ pyamg/classical/gen_air.py
@@ -1,0 +1,305 @@
+"""Approximate ideal restriction AMG"""
+from __future__ import absolute_import
+
+__docformat__ = "restructuredtext en"
+
+from warnings import warn
+from scipy.sparse import csr_matrix, bsr_matrix, isspmatrix_csr, isspmatrix_bsr, \
+    SparseEfficiencyWarning, block_diag, vstack
+from scipy.sparse.linalg import spilu
+import numpy as np
+from copy import deepcopy
+
+from pyamg.multilevel import multilevel_solver
+from pyamg.relaxation.smoothing import change_smoothers
+from pyamg.strength import classical_strength_of_connection, \
+    symmetric_strength_of_connection, evolution_strength_of_connection, \
+    distance_strength_of_connection, algebraic_distance, affinity_distance, \
+    energy_based_strength_of_connection
+from pyamg.util.utils import unpack_arg, extract_diagonal_blocks, \
+    filter_matrix_rows
+from pyamg.classical.interpolate import direct_interpolation, \
+    standard_interpolation, distance_two_interpolation, injection_interpolation, \
+    one_point_interpolation, neumann_AIR, local_AIR
+from pyamg.classical.split import RS, PMIS, PMISc, CLJP, CLJPc, MIS
+from pyamg.classical.cr import CR
+
+__all__ = ['AIR_solver']
+
+def AIR_solver(A,
+               strength=('classical', {'theta': 0.3 ,'norm': 'min'}),
+               CF=('RS', {'second_pass': True}),
+               interp={'P_theta': 0.0, 'Minv' = 'diag'},
+               restrict=('air', {'theta': 0.05, 'degree': 2}),
+               presmoother=None,
+               postsmoother=('FC_jacobi', {'omega': 1.0, 'iterations': 1,
+                              'withrho': False,  'F_iterations': 2,
+                              'C_iterations': 0} ),
+               filter_operator=None,
+               max_levels=20, max_coarse=20,
+               keep=False, **kwargs):
+    """Create a multilevel solver using Classical AMG (Ruge-Stuben AMG)
+
+    Parameters
+    ----------
+    A : csr_matrix
+        Square nonsymmetric matrix in CSR format
+    strength : ['symmetric', 'classical', 'evolution', 'distance',
+                'algebraic_distance','affinity', 'energy_based', None]
+        Method used to determine the strength of connection between unknowns
+        of the linear system.  Method-specific parameters may be passed in
+        using a tuple, e.g. strength=('symmetric',{'theta' : 0.25 }). If
+        strength=None, all nonzero entries of the matrix are considered strong.
+    CF : {string} : default 'RS'
+        Method used for coarse grid selection (C/F splitting)
+        Supported methods are RS, PMIS, PMISc, CLJP, CLJPc, and CR.
+    interp : dictionary with interpolation options:
+        ~ P_theta : {float} default 0.0; Filter tolerance for "strong"
+        connections in Afc for forming P
+        ~ Minv : {string} approximation to Aff^{-1}, options diag and ilu
+    restrict : {string} : default distance-2 AIR, with theta = 0.05.
+        Options include 'air' for local approximate ideal restriction (lAIR)
+        and 'neumann' for Neumann approximate ideal restriction (nAIR).
+    presmoother : {string or dict} : default None
+        Method used for presmoothing at each level.  Method-specific parameters
+        may be passed in using a tuple, e.g.
[…]
