### 2022-04-13 #4142 @mattmartineau
From my perspective the project is great and I would recommend it is accepted.

To summarise - this is an interesting area and the code exposes many novel methods and options, with a good code structure, and clear and detailed documentation. The paper is convincing regarding the necessity of the project and demonstrates clearly the fundamental design, as well as describing how to run a test problem. This also matched my experience of testing the code directly.

The project met all of the objective criteria described in the checklist above.
