### 2022-04-13 #4142 @jedbrown
Great, thanks. @lukeolson I only notice very minor stuff in the paper:
* The citations at the end of the first paragraph are phrased as "in-text", and thus should use `@key` instead of `[@key]`; see https://pandoc.org/MANUAL.html#extension-citations; this is similar to `\citet` versus `\citep`.
* One bib entry should probably protect capitalization `{MueLu}`.

With that, could you:
