+++ pyproject.toml
@@ -5,7 +5,7 @@ requires = [
     "setuptools_scm[toml]>=7.0.0",
     "pybind11>=2.8.0",
     "wheel",
-    "extension-helpers",
+    "extension-helpers@git+https://github.com/gdurif/extension-helpers.git@support_apple_clang",
 ]
 
 [tool.setuptools_scm]
