+++ docs/paper/paper.md
@@ -48,7 +48,7 @@ conditioning of the linear system and convergence rate of the iterative method
 are both dramatically improved.
 `PyAMG` constructs multigrid solvers for use as a
 preconditioner in this setting.  A summary of multigrid and algebraic multigrid
-solvers can be found in @encamg and [-@encmg], and in @amgintro; a detailed description can be found
+solvers can be found in Olson [-@encamg,-@encmg] and in @amgintro; a detailed description can be found
 in @mgtutorial and @mgbook.
 
 # Summary
