### 2022-04-13 fd74188
revert to basic author-in-text
