+++ docs/paper/paper.bib
@@ -138,8 +138,8 @@
 }
 
 @manual{muelu-website,
-  title  = {The {M}uelu {P}roject {W}ebsite},
-  author = {The {M}uelu {P}roject {T}eam},
+  title  = {The {M}ue{L}u {P}roject {W}ebsite},
+  author = {The {M}ue{L}u {P}roject {T}eam},
   year   = {2020},
   url    = {https://trilinos.github.io/muelu.html}
 }
