### 2022-04-13 #4142 @lukeolson
> Great, thanks. @lukeolson I only notice very minor stuff in the paper:
>
> * The citations at the end of the first paragraph are phrased as "in-text", and thus should use `@key` instead of `[@key]`; see https://pandoc.org/MANUAL.html#extension-citations; this is similar to `\citet` versus `\citep`.
> * One bib entry should probably protect capitalization `{MueLu}`.

Thanks @jedbrown.  Fixed the author-in-text (not ideal, I didn't see how to end up with " in Olson (2015a,2015b)", but I'm fine with it.). And fixed the bib.

I'll move on to tagging and loop back.

>
> With that, could you:
>
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
> * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
> * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
> * [ ]  Please list the DOI of the archived version here.
>
> I can then move forward with accepting the submission.
