+++ README-DEV.md
@@ -8,6 +8,7 @@ Release Steps:
 - remove untracked files `git clean -xdf`
 - the following can be done with a pre-release, `v4.2.3-alpha.6`, for testing.  It will not become the default on pypi and `gh release create` can be marked with `--prerelease` (below)
 - mark `fallback_version` in `pyproject.toml`
+- commit, push
 - `git tag -a v3.2.0 -m "version 3.2.0"`
 - `git push`
 - `git push --tags`
