+++ docs/paper/paper.bib
@@ -117,7 +117,7 @@
   year     = {2002},
   note     = {Developments and Trends in Iterative Methods for Large Systems of Equations - in memorium Rudiger Weiss},
   issn     = {0168-9274},
-  doi      = {https://doi.org/10.1016/S0168-9274(01)00115-5},
+  doi      = {10.1016/S0168-9274(01)00115-5},
   url      = {https://www.sciencedirect.com/science/article/pii/S0168927401001155},
   author   = {Van Emden Henson and Ulrike Meier Yang},
   keywords = {Algebraic multigrid, Parallel computing},
