### 2022-04-16 #4142 @lukeolson
@jedbrown fixed (removed the `https://doi.org/`).  Do I need to tag a new release now (and new zenodo doi)?
